PipeOpProxy

Wraps another PipeOp or Graph as determined by the content hyperparameter. Input is routed through the content and the contents' output is returned. The content hyperparameter can be changed during tuning, this is useful as an alternative to PipeOpBranch.

Format

Abstract R6Class inheriting from PipeOp.

Construction

PipeOpProxy$new(innum = 0, outnum = 1, id = "proxy", param_vals = list())

innum :: numeric(1)\cr Determines the number of input channels. If innum` is 0 (default), a vararg input channel is created that can take an arbitrary number of inputs.

outnum :: `numeric(1)
Determines the number of output channels.

id :: character(1)
Identifier of resulting object. See $id slot of PipeOp.

param_vals :: named list
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default list().

Input and Output Channels

PipeOpProxy has multiple input channels depending on the innum construction argument, named "input1", "input2", ... if innum is nonzero; if innum is 0, there is only one vararg input channel named "...".

PipeOpProxy has multiple output channels depending on the outnum construction argument, named "output1", "output2", ... The output is determined by the output of the content operation (a PipeOp or Graph).

State

The $state is the trained content PipeOp or Graph.

Parameters

content :: PipeOp | Graph
The PipeOp or Graph that is being proxied (or an object that is converted to a Graph by as_graph()). Defaults to an instance of PipeOpFeatureUnion (combines all input if they are Tasks).

Internals

The content will internally be coerced to a graph via as_graph() prior to train and predict.

The default value for content is PipeOpFeatureUnion,

Fields

Fields inherited from PipeOp.

Methods

Only methods inherited from PipeOp.

See also

Examples

library("mlr3")

task1 = tsk("iris")
task2 = task1$clone(deep = TRUE)

pop = po("pca", param_vals = list(scale. = TRUE))
pop$train(list(task1))

#> $output
#> <TaskClassif:iris> (150 x 5)
#> * Target: Species
#> * Properties: multiclass
#> * Features (4):
#>   - dbl (4): PC1, PC2, PC3, PC4
#> 

pop$state

#> Standard deviations (1, .., p=4):
#> [1] 1.7083611 0.9560494 0.3830886 0.1439265
#> 
#> Rotation (n x k) = (4 x 4):
#>                     PC1         PC2        PC3        PC4
#> Petal.Length  0.5804131 -0.02449161  0.1421264 -0.8014492
#> Petal.Width   0.5648565 -0.06694199  0.6342727  0.5235971
#> Sepal.Length  0.5210659 -0.37741762 -0.7195664  0.2612863
#> Sepal.Width  -0.2693474 -0.92329566  0.2443818 -0.1235096

proxy = po("proxy", innum = 1, outnum = 1,
  param_vals = list(content = po("pca", param_vals = list(scale. = TRUE))))
proxy$train(list(task2))

#> $output1
#> <TaskClassif:iris> (150 x 5)
#> * Target: Species
#> * Properties: multiclass
#> * Features (4):
#>   - dbl (4): PC1, PC2, PC3, PC4
#> 

proxy$state$state

#> NULL