PipeOpTaskPreproc

Base class for handling most "preprocessing" operations. These are operations that have exactly one Task input and one Task output, and expect the column layout of these tasks during input and output to be the same.

Users must implement $train_task() and $predict_task(), which have a Task input and should return that Task. The Task should, if possible, be manipulated in-place, and should not be cloned.

If can_subset_cols is TRUE, then the hyperparameter affect_columns is added, which is a function that takes the Task as input and returns a character as output, indicating what columns to choose.

PipeOpTaskPreproc

Format

Abstract R6Class inheriting from PipeOp.

Public Members / Active Bindings

intasklayout :: data.table with columns id, type
Set during $train(): $feature_types of training Task, is used for verification during $predict() that columns have not changed.

outtasklayout :: data.table with columns id, type
Set during $train(): $feature_types of transformed Task, is used for verification during $predict() that the transformed Task is the same as during training.

affect_columns :: NULL | function
Can be used to subset the columns that the PipeOp operates on. If $affect_columns is NULL, then all columns are given to the $train / $predict function. If this is a function, it should take one argument (the input Task) and return a character indicating the feature columns that should be operated on. $affect_columns() is then called during training and only features named in its return value are seen by the training / prediction function.
This should not be changed if can_subset_columns is FALSE, and depends on the PipeOpTaskPreproc subclass.

affected_cols :: character
Set during $train(): Indicating the names of features to operate on, if parameter $affect_columns is set.

can_subset_columns :: logical(1)
Whether affect_columns can be set to something other than NULL. Read-only, value depends on the PipeOpTaskPreproc subclass.

Methods

PipeOpTaskPreproc$new(id, param_set = ParamSet$new(), can_subset_cols = TRUE)
(character(1), ParamSet, logical(1)) -> self
Constructor. The can_subset_cols argument should be set by the inheriting class to FALSE if affecting only a subset of columns chosen by the user should not be possible; this switches can_subset_cols to FALSE.

train_task(task)
(Task) -> Task
Train PipeOpTaskPreproc on task, transform it and store a state in $state. The transformed Task must be returned. task should not be cloned, instead it should be changed in-place.
This method can be overloaded when inheriting from PipeOpTaskPreproc, together with $predict_task(); alternatively, $train_dt() and $predict_dt() (and possibly $select_cols()) can be overloaded.

predict_task(task)
(Task) -> Task
Predict on new data in task, possibly using the stored $state. task should not be cloned, instead it should be changed in-place.
This method can be overloaded when inheriting from PipeOpTaskPreproc, together with $train_task(); alternatively, $train_dt() and $predict_dt() (and possibly $select_cols()) can be overloaded.

train_dt(dt)
(data.table) -> any
Train PipeOpTaskPreproc on dt, transform it and store a state in $state. A transformed object must be returned that can be converted to a data.table using as.data.table. dt does not need to be copied deliberately, it is possible and encouraged to change it in-place.
This method can be overloaded when inheriting PipeOpTaskPreproc, together with $predict_dt() and optionally $select_cols(); alternatively, $train_task() and $predict_task() can be overloaded.

predict_dt(dt)
(data.table) -> any
Predict on new data in dt, possibly using the stored $state. A transformed object must be returned that can be converted to a data.table using as.data.table. dt does not need to be copied deliberately, it is possible and encouraged to change it in-place.
This method can be overloaded when inheriting PipeOpTaskPreproc, together with $train_dt() and optionally $select_cols(); alternatively, $train_task() and $predict_task() can be overloaded.

select_cols(dt)
(data.table) -> character
Selects which columns the PipeOp operates on, if $train_dt() and $predict_dt() are overloaded. This function is not called if $train_task() and $predict_task() are overloaded. It is furthermore not supposed to be changed by the user, who should use affect_columns. select_cols is for the ineriting class to determine which columns the operator should function on, e.g. based on feature type.
This method can optionally be overloaded when inheriting PipeOpTaskPreproc, together with $train_dt() and $predict_dt(); alternatively, $train_task() and $predict_task() can be overloaded.
If this method is not overloaded, it defaults to selecting all columns.

See also