PipeOpTaskPreprocSimple
Source: R/PipeOpTaskPreproc.R

Base class for handling many "preprocessing" operations that that perform essentially the same operation during training and prediction. Instead implementing a train_task and a predict_task operation, only a get_state() and a transform() operation needs to be defined, both of which take one argument: a Task.

get_state must not change its input value in-place and must return something that will be written into $state (which must not be NULL), transform() must modify its argument in-state; it is called both during training and prediction.

This inherits from PipeOpTaskPreproc and behaves essentially the same.

PipeOpTaskPreprocSimple

Format

Abstract R6Class inheriting from PipeOpTaskPreproc.

Methods

get_state(task)
(Task) -> any
Store create something that will be stored in $state during training phase of PipeOpTaskPreprocSimple. The state can then influence the $transform() function. Note that $get_state() must return the state, and should not store it in $state. If neither $get_state() nor $get_state_dt() are overloaded, the state will be stored as list().
This method can optionally be overloaded when inheriting from PipeOpTaskPreprocSimple, together with $transform(); alternatively, $get_state_dt() (optional) and $transform_dt() (and possibly $select_cols(), from PipeOpTaskPreproc) can be overloaded.

transform(task)
(Task) -> Task
Predict on new data in task, possibly using the stored $state. task should not be cloned, instead it should be changed in-place. This method is called both during training and prediction phase, and should essentially behave the same independently of phase. (If this is incongruent with the design of the PipeOp, then it should inherit from PipeOpTaskPreproc, not from PipeOpTaskPreprocSimple.)
This method can be overloaded when inheriting from PipeOpTaskPreprocSimple, optionally with $get_state(); alternatively, $get_state_dt() (optional) and $transform_dt() (and possibly $select_cols(), from PipeOpTaskPreproc) can be overloaded.

get_state_dt(dt)
Store create something that will be stored in $state during training phase of PipeOpTaskPreprocSimple. The state can then influence the $transform_dt() function. Note that $get_state_dt() must return the state, and should not store it in $state. If neither $get_state() nor $get_state_dt() are overloaded, the state will be stored as list().
This method can optionally be overloaded when inheriting from PipeOpTaskPreprocSimple, together with $transform_dt() (and optionally $select_cols(), from PipeOpTaskPreproc); Alternatively, $get_state() (optional) and $transform() can be overloaded.

transform_dt(dt)
(data.table) -> any
Predict on new data in dt, possibly using the stored $state. A transformed object must be returned that can be converted to a data.table using as.data.table. dt does not need to be copied deliberately, it is possible and encouraged to change it in-place. This method is called both during training and prediction phase, and should essentially behave the same independently of phase. (If this is incongruent with the design of the PipeOp, then it should inherit from PipeOpTaskPreproc, not from PipeOpTaskPreprocSimple.)
This method can optionally be overloaded when inheriting from PipeOpTaskPreprocSimple, together with $transform_dt() (and optionally $select_cols(), from PipeOpTaskPreproc); Alternatively, $get_state() (optional) and $transform() can be overloaded.

See also