Conversion to mlr3pipeline Graph
Source: R/assert_graph.R

The object is turned into a Graph if possible. If clone is TRUE, a deep copy is made if the incoming object is a Graph to ensure the resulting object is a different reference from the incoming object.

as_graph(x, clone = FALSE)

Arguments

x
(any)
Object to convert.

clone
(logical(1))
Whether to return a (deep copied) clone if x is a Graph.

Value

Graph x or a deep clone of it.

See also