PipeOp Composition Operator

This operator “pipes” data from the source g1 into the sink g2. Both source and sink can either be a Graph or a PipeOp. The number of output channels of g1 (as given by g1$output) must equal the number of input channels of g2 (as given by g2$input). Connections between channels are created in the order in which they occur in g1 and g2, respectively: g1's output channel 1 is connected to g2's input channel 1, channel 2 to 2 etc.

This operator always created deep copies of its input arguments, so they cannot be modified by reference. To access individual PipeOps after composition, use the resulting Graph's $pipeops list.

g1 %>>% g2

Arguments

g1
(Graph | PipeOp)
Graph / PipeOp to put in front of g2.

g2
(Graph | PipeOp)
Graph / PipeOp to put after g1.

Value

Graph the constructed Graph.

See also

Examples

o1 = PipeOpScale$new()
o2 = PipeOpPCA$new()

# The following two are equivalent:
result1 = o1 %>>% o2

result2 = Graph$new()$
  add_pipeop(o1$clone(deep = TRUE))$
  add_pipeop(o2$clone(deep = TRUE))$
  add_edge(o1$id, o2$id)