Dictionary of (sub-)graphs

A simple Dictionary storing objects of class Graph. The dictionary contains a collection of often-used graph structures, and it's aim is solely to make often-used functions more accessible. Each Graph has an associated help page, which can be accessed via ?mlr_graphs_<key>, i.e. ?mlr_graphs_bagging.

Format

R6Class object inheriting from mlr3misc::Dictionary.

Methods

Methods inherited from Dictionary, as well as:

add(key, value)
(character(1), function)
Adds constructor value to the dictionary with key key, potentially overwriting a previously stored item.

S3 methods

as.data.table(dict)
Dictionary -> data.table::data.table
Returns a data.table with column key (character).

See also

Examples

library(mlr3)
lrn = lrn("regr.rpart")
task = mlr_tasks$get("boston_housing")

# Robustify the learner for the task.
gr = pipeline_robustify(task, lrn) %>>% po("learner", lrn)
# or equivalently
gr = mlr_graphs$get("robustify", task = task, learner = lrn) %>>% po(lrn)
# or equivalently
gr = ppl("robustify", task, lrn) %>>% po("learner", lrn)

# all Graphs currently in the dictionary:
as.data.table(mlr_graphs)
#> Key: <key>
#>              key
#>           <char>
#> 1:       bagging
#> 2:        branch
#> 3: convert_types
#> 4:    greplicate
#> 5:           ovr
#> 6:     robustify
#> 7:      stacking
#> 8:   targettrafo