PipeOpBranch

This pipeop is used for multiplexing between different possible paths and should be used in conjunction with PipeOpUnbranch.

PipeOpBranch

Format

R6Class object inheriting from PipeOp.

Methods

PipeOpBranch$new(options, id = "branch")
(numeric(1) | character, character(1)) -> self
Constructor. If options is an integer number, it determines the number of output channels / options that are created, named output1...output<n>. The $selection parameter will then be a ParamInt. If options is a character, it determines the names of channels directly. The $selection parameter will then be a ParamFct.

Parameters

selection: (numeric(1) | character(1))
Selection of branching path to take. Is a ParamInt if the options parameter during construction was a numeric(1), and ranges from 1 to options. Is a ParamFct if the options parameter was a character and its possible values are the options values.

Details

Creates a PipeOp with multiple output channels that can be used to create a Graph network with alternative paths. If options is an integer(1), it determines the number of out-paths and selection is an integer parameter choosing between these paths. If options is a character, then length(options) out channels are created, each named according to options.

To create a usable graph, the branching paths need to be brought together using PipeOpUnbranch.

Not to be confused with PipeOpCopy, the naming scheme is a bit unfortunate.

See also

Examples

pca = PipeOpPCA$new()
nop = PipeOpNULL$new()
choices = c("pca", "nothing")
PipeOpBranch$new(choices) %>>% gunion(list(pca, nop)) %>>% PipeOpUnbranch$new(choices)

#> Graph with 4 PipeOps:
#>        ID         State sccssors prdcssors
#>    branch <<UNTRAINED>> pca,null          
#>       pca <<UNTRAINED>> unbranch    branch
#>      null <<UNTRAINED>> unbranch    branch
#>  unbranch <<UNTRAINED>>           pca,null