PipeOpEncode
Source: R/PipeOpEncode.R

Encodes factor, character and ordered factor columns.

Possible encodings are "one-hot" encoding, as well as encoding according to stats::contr.helmert(), stats::contr.poly(), stats::contr.sum() and stats::contr.treatment(). Newly created columns are named by factor levels in "one-hot" and "treatment" encoding, and numbered otherwise.

Use the $affect_columns functionality to only encode a subset of columns, or only encode columns of a certain type.

PipeOpEncode

Format

R6Class object inheriting from PipeOpTaskPreprocSimple.

Parameter Set

method :: character(1)
If set to "one-hot", creates a new column for each factor level. If set to "treatment", creates $n-1$ columns leaving out the first factor level of each factor variable (see stats::contr.treatment()). If set to "helmert", creates columns according to Helmert contrasts (see stats::contr.helmert()). If set to "poly", creates columns with contrasts based on orthogonal polynomials (see stats::contr.poly()). If set to "sum", creates columns with contrasts summing to zero, (see stats::contr.sum()). `

See also