PipeOpFilter
Source: R/PipeOpFilter.R

Feature filtering using a mlr3featsel::Filter object, see the mlr3featsel package.

The settings of the filter are given as paradox::ParamUty; when Filters start supporting the paradox interface that could be used instead.

If a Filter can only operate on a subset of columns based on column type, then only these features are considered. nfeat and frac will count for the features of the type that the Filter can operate on; this means e.g. that setting nfeat to 0 will only remove features of the type that the Filter can work with.

PipeOpFilter

Format

R6Class object inheriting from PipeOpTaskPreprocSimple.

Methods

PipeOpFilter$new(filter, id = filter$id, param_vals = list())
(mlr3featsel::Filter, character(1), list) -> self
Constructor. filter gives the Filter to use.

Parameter Set

settings :: named list
List of settings to be given to the Filter. Default list().

nfeat :: numeric(1)
Number of features to select. If this is set, frac and cutoff` must not be set.

frac :: numeric(1)
Fraction of features to keep. If this is set, nfeat and cutoff must not be set.

cutoff :: numeric(1)
Minimum value of filter heuristic for which to keep features. If this is set, nfeat and frac must not be set.

See also