PipeOpLearnerCV
Source: R/PipeOpLearnerCV.R

Wraps an mlr3::Learner into a PipeOp. Returns cross-validated predictions during training phase and stores a model of the Learner trained on the whole data in $state as a new Task with the predictions; usually a feature <ID>.response if the Learner's $predict.type is "response", <ID>.prob.<CLASS> and <ID>.response for $predict.type "prob", or <ID>.response and <ID>.se for $predict.type "se". <ID> is here the $id of the PipeOpLearnerCV.

Returns this model's prediction during prediction phase, as a new Task with a single column.

Inherits the $param_set and therefore $param_set$values from the Learner it is constructed from.

PipeOpLearnerCV

Format

R6Class object inheriting from PipeOpTaskPreproc.

Public Members / Active Bindings

learner :: Learner
Learner to use for cross validation / prediction.

Methods

PipeOpLearner$new(learner, id = learner$id)
(Learner, character(1)) -> self
Constructor. The given learner will be used for crossvalidation.

Parameter Set

resampling :: character(1)
Which resampling method do we want to use. Currently only supports 'cv'.

folds :: numeric(1)
Number of cross validation folds.

...
The ParamSet of $learner is also made available.

See also