PipeOpUnbranch
Source: R/PipeOpUnBranch.R

Used to bring together different paths created by PipeOpBranch.

PipeOpUnbranch

Format

R6Class object inheriting from PipeOp.

Methods

PipeOpUnbranch$new(options, id = "unbranch")
(numeric(1) | character, character(1)) -> self
Constructor. If options is an integer number, it determines the number of input channels that are created, named input1...input<n>. If options is a character, it determines the names of channels directly.

Details

Creates a PipeOp with multiple input channels that can be used to create a Graph network with alternative paths. options works as in PipeOpBranch and should probably be the same value as the options given to the corresponding PipeOpBranch instance.

See also

Examples

pca = PipeOpPCA$new()
nop = PipeOpNULL$new()
choices = c("pca", "nothing")
PipeOpBranch$new(choices) %>>% gunion(list(pca, nop)) %>>% PipeOpUnbranch$new(choices)

#> Graph with 4 PipeOps:
#>        ID         State sccssors prdcssors
#>    branch <<UNTRAINED>> pca,null          
#>       pca <<UNTRAINED>> unbranch    branch
#>      null <<UNTRAINED>> unbranch    branch
#>  unbranch <<UNTRAINED>>           pca,null