Dictionary of PipeOps
Source: R/mlr_pipeops.R

A simple Dictionary storing objects of class PipeOp. Each PipeOp has an associated help page, see mlr_pipeops_[id].

Format

R6Class object inheriting from mlr3misc::Dictionary.

Fields

Fields inherited from Dictionary, as well as:

metainf :: environment
Environment that stores the metainf argument of the $add() method. Only for internal use.

Methods

Methods inherited from Dictionary, as well as:

add(key, value, metainf = NULL)
(character(1), R6ClassGenerator, NULL | list)
Adds constructor value to the dictionary with key key, potentially overwriting a previously stored item. If metainf is not NULL (the default), it must be a list of arguments that will be given to the value constructor (i.e. value$new()) when it needs to be constructed for as.data.table PipeOp listing.

S3 methods

as.data.table(dict)
Dictionary -> data.table::data.table
Returns a data.table with columns key (character), packages (character), input.num (integer), output.num (integer), input.type.train (character), input.type.predict (character), output.type.train (character), output.type.predict (character).

See also

Examples

library(mlr3)

mlr_pipeops$get("learner", lrn("classif.rpart"))

#> PipeOp: <classif.rpart> (not trained)
#> values: <xval=0>
#> Input channels <name [train type, predict type]>:
#>   input [TaskClassif,TaskClassif]
#> Output channels <name [train type, predict type]>:
#>   output [NULL,PredictionClassif]

# equivalent:
po("learner", learner = lrn("classif.rpart"))

#> PipeOp: <classif.rpart> (not trained)
#> values: <xval=0>
#> Input channels <name [train type, predict type]>:
#>   input [TaskClassif,TaskClassif]
#> Output channels <name [train type, predict type]>:
#>   output [NULL,PredictionClassif]