PipeOpFeatureUnion
Source: R/PipeOpFeatureUnion.R

Aggregates features from all input tasks by cbind()ing them together into a single Task.

DataBackend primary keys and Task targets have to be equal across all Tasks. Only the target column(s) of the first Task are kept.

If assert_targets_equal is TRUE then target column names are compared and an error is thrown if they differ across inputs.

Format

R6Class object inheriting from PipeOp.

Construction

PipeOpFeatureUnion$new(innum = 0, id = "featureunion", param_vals = list(), assert_targets_equal = TRUE)

innum :: numeric(1) | character
Determines the number of input channels. If innum is 0 (default), a vararg input channel is created that can take an arbitrary number of inputs. If innum is a character vector, the number of input channels is the length of innum, and the columns of the result are prefixed with the values.

id :: character(1) Identifier of the resulting object, default "featureunion".

param_vals :: named list
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default list().

assert_targets_equal :: logical(1)
If assert_targets_equal is TRUE (Default), task target column names are checked for agreement. Disagreeing target column names are usually a bug, so this should often be left at the default.

Input and Output Channels

PipeOpFeatureUnion has multiple input channels depending on the innum construction argument, named "input1", "input2", ... if innum is nonzero; if innum is 0, there is only one vararg input channel named "...". All input channels take a Task both during training and prediction.

PipeOpFeatureUnion has one output channel named "output", producing a Task both during training and prediction.

The output is a Task constructed by cbind()ing all features from all input Tasks, both during training and prediction.

State

The $state is left empty (list()).

Parameters

PipeOpFeatureUnion has no Parameters.

Internals

PipeOpFeatureUnion uses the Task $cbind() method to bind the input values beyond the first input to the first Task. This means if the Tasks are database-backed, all of them except the first will be fetched into R memory for this. This behaviour may change in the future.

Fields

Only fields inherited from PipeOp.

Methods

Only methods inherited from PipeOp.

See also

Examples

library(mlr3)

task = tsk("iris")
gr = gunion(list(
  po("nop"),
  po("pca")
)) %>>% po("featureunion")

gr$train(task)

#> $featureunion.output
#> <TaskClassif:iris> (150 x 9)
#> * Target: Species
#> * Properties: multiclass
#> * Features (8):
#>   - dbl (8): PC1, PC2, PC3, PC4, Petal.Length, Petal.Width,
#>     Sepal.Length, Sepal.Width
#> 

po = po("featureunion", innum = c("a", "b"))

po$train(list(task, task))

#> [[1]]
#> <TaskClassif:iris> (150 x 9)
#> * Target: Species
#> * Properties: multiclass
#> * Features (8):
#>   - dbl (8): a.Petal.Length, a.Petal.Width, a.Sepal.Length,
#>     a.Sepal.Width, b.a.Petal.Length, b.a.Petal.Width, b.a.Sepal.Length,
#>     b.a.Sepal.Width
#>