PipeOpFilter
Source: R/PipeOpFilter.R

Feature filtering using a mlr3filters::Filter object, see the mlr3filters package.

If a Filter can only operate on a subset of columns based on column type, then only these features are considered and filtered. nfeat and frac will count for the features of the type that the Filter can operate on; this means e.g. that setting nfeat to 0 will only remove features of the type that the Filter can work with.

Format

R6Class object inheriting from PipeOpTaskPreprocSimple/PipeOpTaskPreproc/PipeOp.

Construction

PipeOpFilter$new(filter, id = filter$id, param_vals = list())

filter :: Filter
Filter used for feature filtering.

id :: character(1) Identifier of the resulting object, defaulting to the id of the Filter being used.

param_vals :: named list
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default list().

Input and Output Channels

Input and output channels are inherited from PipeOpTaskPreproc.

The output is the input Task with features removed that were filtered out.

State

The $state is a named list with the $state elements inherited from PipeOpTaskPreproc, as well as:

scores :: named numeric
Scores calculated for all features of the training Task which are being used as cutoff for feature filtering. If frac or nfeat is given, the underlying Filter may choose to not calculate scores for all features that are given. This only includes features on which the Filter can operate; e.g. if the Filter can only operate on numeric features, then scores for factorial features will not be given.

features :: character
Names of features that are being kept. Features of types that the Filter can not operate on are always being kept.

Parameters

The parameters are the parameters inherited from the PipeOpTaskPreproc, as well as the parameters of the Filter used by this object. Besides, parameters introduced are:

filter.nfeat :: numeric(1)
Number of features to select. Mutually exclusive with frac and cutoff.

filter.frac :: numeric(1)
Fraction of features to keep. Mutually exclusive with nfeat and cutoff.

filter.cutoff :: numeric(1)
Minimum value of filter heuristic for which to keep features. Mutually exclusive with nfeat and frac.

Note that at least one of filter.nfeat, filter.frac, or filter.cutoff must be given.

Internals

This does not use the $select_cols feature of PipeOpTaskPreproc to select only features compatible with the Filter; instead the whole Task is used by $get_state() and subset internally.

Fields

Fields inherited from PipeOpTaskPreproc, as well as:

filter :: Filter
Filter that is being used for feature filtering. Do not use this slot to get to the feature filtering scores after training; instead, use $state$scores. Read-only.

Methods

Methods inherited from PipeOpTaskPreprocSimple/PipeOpTaskPreproc/PipeOp.

See also

Examples

library(mlr3)
library(mlr3filters)

# setup PipeOpFilter to keep the 5 most important
# features of the spam task w.r.t. their AUC
task = tsk("spam")
filter = flt("auc")
po = po("filter", filter = filter)
po$param_set

#> ParamSet: auc
#>                id    class lower upper levels     default value
#> 1:   filter.nfeat ParamInt     0   Inf        <NoDefault>      
#> 2:    filter.frac ParamDbl     0     1        <NoDefault>      
#> 3:  filter.cutoff ParamDbl  -Inf   Inf        <NoDefault>      
#> 4: affect_columns ParamUty    NA    NA        <NoDefault>      

po$param_set$values$filter.nfeat = 5

# filter the task
filtered_task = po$train(list(task))[[1]]

# filtered task + extracted AUC scores
filtered_task$feature_names

#> [1] "capitalAve"      "capitalLong"     "charDollar"      "charExclamation"
#> [5] "your"           

head(po$state$scores, 10)

#> charExclamation     capitalLong      capitalAve            your      charDollar 
#>       0.3290461       0.3041626       0.2882004       0.2801659       0.2721394 
#>    capitalTotal            free             our             you          remove 
#>       0.2622801       0.2327285       0.2109325       0.2104681       0.2031303 

# feature selection embedded in a 3-fold cross validation
# keep 30% of features based on their AUC score
task = tsk("spam")
gr = po("filter", filter = flt("auc"), filter.frac = 0.5) %>>%
  po("learner", lrn("classif.rpart"))
learner = GraphLearner$new(gr)
rr = resample(task, learner, rsmp("cv", folds = 3), store_models = TRUE)

#> INFO  [14:12:33.970] Applying learner 'auc.classif.rpart' on task 'spam' (iter 1/3) 
#> INFO  [14:12:34.101] Applying learner 'auc.classif.rpart' on task 'spam' (iter 2/3) 
#> INFO  [14:12:34.245] Applying learner 'auc.classif.rpart' on task 'spam' (iter 3/3) 

rr$learners[[1]]$model$auc$scores

#>   charExclamation       capitalLong        capitalAve              your 
#>       0.335219380       0.309310614       0.297876272       0.272630111 
#>        charDollar      capitalTotal              free               you 
#>       0.268664747       0.262582575       0.237378370       0.211729379 
#>               our            remove                hp             money 
#>       0.210005133       0.202402026       0.175193046       0.171419389 
#>               all            num000          business              mail 
#>       0.167199830       0.154013346       0.143544456       0.137876495 
#>            george               hpl          internet              over 
#>       0.137235315       0.132019729       0.131953669       0.131558650 
#>           receive             email           address             order 
#>       0.125929075       0.125424031       0.118521469       0.111289056 
#>          charHash              make           num1999            credit 
#>       0.103226433       0.098824987       0.097871585       0.090299946 
#>              will            people              labs         addresses 
#>       0.084735538       0.073679700       0.072197822       0.068600473 
#>            num650               edu             num85               lab 
#>       0.068045661       0.066362034       0.064964069       0.060488529 
#>        technology           meeting              data            telnet 
#>       0.051917292       0.051743662       0.047165238       0.046583423 
#>                pm           project            report            num857 
#>       0.041750357       0.041452419       0.038991698       0.031996742 
#> charSquarebracket            num415        conference                re 
#>       0.031573826       0.030525576       0.030341903       0.026796777 
#>          original                cs     charSemicolon              font 
#>       0.025873505       0.023971389       0.023677468       0.022070836 
#>  charRoundbracket            direct             num3d             parts 
#>       0.014496295       0.013620559       0.009517051       0.003412114 
#>             table 
#>       0.002961525