PipeOpImpute
Source: R/PipeOpImpute.R

Impute missing values using varying methods.

numeric or integer features are imputed by method_num.

factor, ordered, and character features are imported by method_fct.

logical features are always imputed by sampling from the training column.

Format

R6Class object inheriting from PipeOpTaskPreprocSimple/PipeOpTaskPreproc/PipeOp.

Construction

PipeOpImpute$new(id = "impute", param_vals = list())

id :: character(1) Identifier of the resulting object, defaulting to "impute".

param_vals :: named list
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default list().

State

$state is a named list with the $state elements inherited from PipeOpTaskPreproc, as well as:

TO BE DESCRIBED ( PipeOp will be changed shortly )

Parameters

The parameters are the parameters inherited from the PipeOpTaskPreproc, as well as:

TO BE DESCRIBED ( PipeOp will be changed shortly )

Internals

Will be refactored shortly.

Fields

Fields inherited from PipeOpTaskPreproc/PipeOp.

Methods

Methods inherited from PipeOpTaskPreproc/PipeOp.

See also

Examples

library(mlr3)

task = tsk("pima")
sum(complete.cases(task$data()))

#> [1] 392

task$missings()

#> diabetes      age  glucose  insulin     mass pedigree pregnant pressure 
#>        0        0        5      374       11        0        0       35 
#>  triceps 
#>      227 

po = po("impute")
new_task = po$train(list(task = task))[[1]]
sum(complete.cases(new_task$data()))

#> [1] 768

new_task$missings()

#>         diabetes              age         pedigree         pregnant 
#>                0                0                0                0 
#>          glucose          insulin             mass         pressure 
#>                0                0                0                0 
#>          triceps  missing_glucose  missing_insulin     missing_mass 
#>                0                0                0                0 
#> missing_pressure  missing_triceps 
#>                0                0