PipeOpImputeMedian

Impute numerical features by their median.

Format

R6Class object inheriting from PipeOpImpute/PipeOp.

Construction

PipeOpImputeMedian$new(id = "imputemedian", param_vals = list())

id :: character(1)
Identifier of resulting object, default "imputemedian".

param_vals :: named list
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default list().

Input and Output Channels

Input and output channels are inherited from PipeOpImputeMedian.

The output is the input Task with all affected numeric features missing values imputed by (column-wise) median.

State

The $state is a named list with the $state elements inherited from PipeOpImpute.

The $state$model is a named list of numeric(1) indicating the median of the respective feature.

Parameters

The parameters are the parameters inherited from PipeOpImpute.

Internals

Uses the stats::median() function. Features that are entirely NA are imputed as 0.

Methods

Only methods inherited from PipeOpImpute/PipeOp.

See also

Examples

library("mlr3")

task = tsk("pima")
task$missings()

#> diabetes      age  glucose  insulin     mass pedigree pregnant pressure 
#>        0        0        5      374       11        0        0       35 
#>  triceps 
#>      227 

po = po("imputemedian")
new_task = po$train(list(task = task))[[1]]
new_task$missings()

#> diabetes      age pedigree pregnant  glucose  insulin     mass pressure 
#>        0        0        0        0        0        0        0        0 
#>  triceps 
#>        0 

po$state$model

#> $age
#> [1] 29
#> 
#> $glucose
#> [1] 117
#> 
#> $insulin
#> [1] 125
#> 
#> $mass
#> [1] 32.3
#> 
#> $pedigree
#> [1] 0.3725
#> 
#> $pregnant
#> [1] 3
#> 
#> $pressure
#> [1] 72
#> 
#> $triceps
#> [1] 29
#>