PipeOpKernelPCA

Extracts kernel principle components from data. Only affects numerical features. See kernlab::kpca for details.

Format

R6Class object inheriting from PipeOpTaskPreproc/PipeOp.

Construction

PipeOpKernelPCA$new(id = "kernelpca", param_vals = list())

id :: character(1)
Identifier of resulting object, default "kernelpca".

param_vals :: named list
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default list().

Input and Output Channels

Input and output channels are inherited from PipeOpTaskPreproc.

The output is the input Task with all affected numeric parameters replaced by their principal components.

State

The $state is a named list with the $state elements inherited from PipeOpTaskPreproc, as well as the returned S4 object of the function kernlab::kpca().

The @rotated slot of the "kpca" object is overwritten with an empty matrix for memory efficiency.

The slots of the S4 object can be accessed by accessor function. See kernlab::kpca.

Parameters

The parameters are the parameters inherited from PipeOpTaskPreproc, as well as:

kernel :: character(1)
The standard deviations of the principal components. See kpca().

kpar :: list
List of hyper-parameters that are used with the kernel function. See kpca().

features :: numeric(1)
Number of principal components to return. Default 0 means that all principal components are returned. See kpca().

th :: numeric(1)
The value of eigenvalue under which principal components are ignored. Default is 0.0001. See kpca().

na.action :: function
Function to specify NA action. Default is na.omit. See kpca().

Internals

Uses the kpca() function.

Methods

Only methods inherited from PipeOpTaskPreproc/PipeOp.

See also

Examples

library("mlr3")

task = tsk("iris")
pop = po("kernelpca", features = 3)  # only keep top 3 components

task$data()

#>        Species Petal.Length Petal.Width Sepal.Length Sepal.Width
#>   1:    setosa          1.4         0.2          5.1         3.5
#>   2:    setosa          1.4         0.2          4.9         3.0
#>   3:    setosa          1.3         0.2          4.7         3.2
#>   4:    setosa          1.5         0.2          4.6         3.1
#>   5:    setosa          1.4         0.2          5.0         3.6
#>  ---                                                            
#> 146: virginica          5.2         2.3          6.7         3.0
#> 147: virginica          5.0         1.9          6.3         2.5
#> 148: virginica          5.2         2.0          6.5         3.0
#> 149: virginica          5.4         2.3          6.2         3.4
#> 150: virginica          5.1         1.8          5.9         3.0

pop$train(list(task))[[1]]$data()

#>        Species        V1         V2           V3
#>   1:    setosa -9.439059 -1.1738319  0.818082979
#>   2:    setosa -9.306342 -0.8132986 -1.523991896
#>   3:    setosa -9.536490 -1.3665450 -1.351184473
#>   4:    setosa -9.279544 -0.8197809 -2.060504832
#>   5:    setosa -9.483230 -1.3216931  0.865439972
#>  ---                                            
#> 146: virginica  6.588363 -2.4365684  0.686844853
#> 147: virginica  6.117199  0.6557377  0.003376105
#> 148: virginica  6.577817 -1.3279651  0.771331382
#> 149: virginica  6.364696 -2.1709066  0.600659654
#> 150: virginica  5.878114  1.0534310  0.323595267