PipeOpMutate
Source: R/PipeOpMutate.R

Adds features according to expressions given as formulas that may depend on values of other features. This can add new features, or can change existing features.

Format

R6Class object inheriting from PipeOpTaskPreprocSimple/PipeOpTaskPreproc/PipeOp.

Construction

PipeOpMutate$new(id = "mutate", param_vals = list())

id :: character(1)
Identifier of resulting object, default "mutate".

param_vals :: named list
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default list().

Input and Output Channels

Input and output channels are inherited from PipeOpTaskPreproc.

The output is the input Task with added and/or mutated features according to the mutation parameter.

State

The $state is a named list with the $state elements inherited from PipeOpTaskPreproc.

Parameters

The parameters are the parameters inherited from PipeOpTaskPreproc, as well as:

mutation :: named list of formula
Expressions for new features to create (or present features to change), in the form of formula. Each element of the list is a formula with the name of the element naming the feature to create or change, and the formula expression determining the result. This expression may reference other features, as well as variables visible at the creation of the formula (see examples). Initialized to list().

delete_originals :: logical(1)
Whether to delete original features. Even when this is FALSE, present features may still be overwritten. Initialized to FALSE.

Internals

A formula created using the ~ operator always contains a reference to the environment in which the formula is created. This makes it possible to use variables in the ~-expressions that both reference either column names or variable names.

Fields

Only fields inherited from PipeOpTaskPreproc/PipeOp.

Methods

Only methods inherited from PipeOpTaskPreprocSimple/PipeOpTaskPreproc/PipeOp.

See also

Examples

library(mlr3)

pom = po("mutate")
pom$param_set$values$mutation = list(
  Sepal.Area = ~ Sepal.Width * Sepal.Length,
  Petal.Area = ~ Petal.Width * Petal.Length
)

pom$train(list(tsk("iris")))[[1]]$data()

#>        Species Petal.Length Petal.Width Sepal.Length Sepal.Width Sepal.Area
#>   1:    setosa          1.4         0.2          5.1         3.5      17.85
#>   2:    setosa          1.4         0.2          4.9         3.0      14.70
#>   3:    setosa          1.3         0.2          4.7         3.2      15.04
#>   4:    setosa          1.5         0.2          4.6         3.1      14.26
#>   5:    setosa          1.4         0.2          5.0         3.6      18.00
#>  ---                                                                       
#> 146: virginica          5.2         2.3          6.7         3.0      20.10
#> 147: virginica          5.0         1.9          6.3         2.5      15.75
#> 148: virginica          5.2         2.0          6.5         3.0      19.50
#> 149: virginica          5.4         2.3          6.2         3.4      21.08
#> 150: virginica          5.1         1.8          5.9         3.0      17.70
#>      Petal.Area
#>   1:       0.28
#>   2:       0.28
#>   3:       0.26
#>   4:       0.30
#>   5:       0.28
#>  ---           
#> 146:      11.96
#> 147:       9.50
#> 148:      10.40
#> 149:      12.42
#> 150:       9.18