PipeOpOVRUnite

Perform "One vs. Rest" classification by (weighted) majority vote prediction from classification Predictions. This works in combination with PipeOpOVRSplit.

Weights can be set as a parameter; if none are provided, defaults to equal weights for each prediction.

Always returns a "prob" prediction, regardless of the incoming Learner's $predict_type. The label of the class with the highest predicted probability is selected as the "response" prediction.

Missing values during prediction are treated as each class label being equally likely.

This PipeOp uses a Multiplicity input, which is created by PipeOpOVRSplit and causes PipeOps on the way to this PipeOp to be called once for each individual binary Task.

Note that Multiplicity is currently an experimental features and the implementation or UI may change.

Format

R6Class inheriting from PipeOpEnsemble/PipeOp.

Construction

PipeOpOVRUnite$new(id = "ovrunite", param_vals = list())

id :: character(1)
Identifier of the resulting object, default "ovrunite".

param_vals :: named list
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default list().

Input and Output Channels

Input and output channels are inherited from PipeOpEnsemble. Instead of a Prediction, a PredictionClassif is used as input and output during prediction and PipeOpEnsemble's collect parameter is initialized with TRUE to allow for collecting a Multiplicity input.

State

The $state is left empty (list()).

Parameters

The parameters are the parameters inherited from the PipeOpEnsemble.

Internals

Inherits from PipeOpEnsemble by implementing the private$.predict() method.

Should be used in combination with PipeOpOVRSplit.

Fields

Only fields inherited from PipeOpEnsemble/PipeOp.

Methods

Only methods inherited from PipeOpEnsemble/PipeOp.

See also

Examples

library(mlr3)
task = tsk("iris")
gr = po("ovrsplit") %>>% lrn("classif.rpart") %>>% po("ovrunite")
gr$train(task)

#> $ovrunite.output
#> NULL
#> 

gr$predict(task)

#> $ovrunite.output
#> <PredictionClassif> for 150 observations:
#>     row_id     truth  response prob.setosa prob.versicolor prob.virginica
#>          1    setosa    setosa           1               0              0
#>          2    setosa    setosa           1               0              0
#>          3    setosa    setosa           1               0              0
#> ---                                                                      
#>        148 virginica virginica           0               0              1
#>        149 virginica virginica           0               0              1
#>        150 virginica virginica           0               0              1
#> 

gr$pipeops$classif.rpart$learner$predict_type = "prob"
gr$predict(task)

#> $ovrunite.output
#> <PredictionClassif> for 150 observations:
#>     row_id     truth  response prob.setosa prob.versicolor prob.virginica
#>          1    setosa    setosa   0.9541284      0.00000000     0.04587156
#>          2    setosa    setosa   0.9541284      0.00000000     0.04587156
#>          3    setosa    setosa   0.9541284      0.00000000     0.04587156
#> ---                                                                      
#>        148 virginica virginica   0.0000000      0.02173913     0.97826087
#>        149 virginica virginica   0.0000000      0.02173913     0.97826087
#>        150 virginica virginica   0.0000000      0.02173913     0.97826087
#>