PipeOpRegrAvg
Source: R/PipeOpRegrAvg.R

Perform (weighted) prediction averaging from regression Predictions by connecting PipeOpRegrAvg to multiple PipeOpLearner outputs.

The resulting "response" prediction is a weighted average of the incoming "response" predictions. "se" prediction is currently not aggregated but discarded if present.

Weights can be set as a parameter; if none are provided, defaults to equal weights for each prediction. Defaults to equal weights for each model.

Format

R6Class inheriting from PipeOpEnsemble/PipeOp.

Construction

PipeOpRegrAvg$new(innum = 0, id = "regravg", param_vals = list())

innum :: numeric(1)
Determines the number of input channels. If innum is 0 (default), a vararg input channel is created that can take an arbitrary number of inputs.

id :: character(1) Identifier of the resulting object, default "regravg".

param_vals :: named list
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default list().

Input and Output Channels

Input and output channels are inherited from PipeOpEnsemble. Instead of a Prediction, a PredictionRegr is used as input and output during prediction.

State

The $state is left empty (list()).

Parameters

The parameters are the parameters inherited from the PipeOpEnsemble.

Internals

Inherits from PipeOpEnsemble by implementing the private$weighted_avg_predictions() method.

Fields

Only fields inherited from PipeOpEnsemble/PipeOp.

Methods

Only methods inherited from PipeOpEnsemble/PipeOp.

See also

Examples

library(mlr3)

# Simple Bagging
gr = greplicate(n = 5,
  po("subsample") %>>%
  po("learner", lrn("classif.rpart"))
) %>>%
  po("classifavg")

resample(tsk("iris"), GraphLearner$new(gr), rsmp("cv", folds = 3))

#> INFO  [14:12:38.903] Applying learner 'subsample_1.subsample_2.subsample_3.subsample_4.subsample_5.classif.rpart_1.classif.rpart_2.classif.rpart_3.classif.rpart_4.classif.rpart_5.classifavg' on task 'iris' (iter 1/3) 
#> INFO  [14:12:39.220] Applying learner 'subsample_1.subsample_2.subsample_3.subsample_4.subsample_5.classif.rpart_1.classif.rpart_2.classif.rpart_3.classif.rpart_4.classif.rpart_5.classifavg' on task 'iris' (iter 2/3) 
#> INFO  [14:12:39.513] Applying learner 'subsample_1.subsample_2.subsample_3.subsample_4.subsample_5.classif.rpart_1.classif.rpart_2.classif.rpart_3.classif.rpart_4.classif.rpart_5.classifavg' on task 'iris' (iter 3/3) 

#> <ResampleResult> of 3 iterations
#> * Task: iris
#> * Learner: subsample_1.subsample_2.subsample_3.subsample_4.subsample_5.classif.rpart_1.classif.rpart_2.classif.rpart_3.classif.rpart_4.classif.rpart_5.classifavg
#> * Performance: 0.053 [classif.ce]
#> * Warnings: 0 in 0 iterations
#> * Errors: 0 in 0 iterations