Transforms Numeric Features into Spline Basis Expansions

Replaces numeric features with columns representing spline basis expansions.

Depending on the type parameter, constructs polynomial B-splines splines::bs() or natural cubic splines splines::ns() for the respective column.

Format

R6Class object inheriting from PipeOpTaskPreproc/PipeOp.

Construction

po("splines", param_vals = list())

id :: character(1)
Identifier of resulting object, default "splines".

param_vals :: named list
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default list().

Input and Output Channels

Input and output channels are inherited from PipeOpTaskPreproc.

The output is the input Task with the selected columns transformed according to the specified Splines Method.

State

The $state is a named list with the $state elements inherited from PipeOpTaskPreproc. After training the Boundary.knots will be given to the $state.

Parameters

The parameters are the parameters inherited from PipeOpTaskPreproc, as well as:

type :: character(1)
Controls the type of splines that are to be created. Can be either polynomial (splines::bs) or natural (splines::ns). Initializied to "natural".

df :: integer(1)
Number of degrees of freedom for calculation of the spline basis matrix. Initialized to NULL. Depending on type, see either splines::bs() or splines::ns().

knots :: named list
Internal breakpoints that define the spline, given as a named list of numeric vectors, where each name corresponds to a feature and its value specifies the knots for that feature. Default is NULL. Depending on type, see either splines::bs() or splines::ns().

intercept :: logical(1)
If TRUE, an intercept is included in the basis. Default is FALSE. Depending on type, see either splines::bs() or splines::ns().

degree :: integer(1)
Degree of the polynomial used to compute polynomial splines. Only used if type is "polynomial". Default is 3. See splines::bs().

Boundary.knots :: named list
Boundary points at which to anchor the spline basis, given as a named list of numeric vectors, where each name corresponds to a feature and its value specifies the boundary points for that feature. Default is NULL. Depending on type, see either splines::bs() or splines::ns().

Internals

Creates a spline basis using either splines::bs or splines::ns depending on the hyperparameter type. After training, the Boundary.knots that were either provided by the user or calculated during training are stored in the PipeOp's $state.

Fields

Only fields inherited from PipeOp.

Methods

Only methods inherited from PipeOpTaskPreproc/PipeOp.

See also

Examples

library("mlr3")

task = tsk("iris")
pop = po("splines")

pop$train(list(task))[[1]]$data()
#>        Species Petal.Length Petal.Width Sepal.Length Sepal.Width
#>         <fctr>        <num>       <num>        <num>       <num>
#>   1:    setosa   0.05435822  0.03340766   0.17817416   0.5011148
#>   2:    setosa   0.05435822  0.03340766   0.13363062   0.3340766
#>   3:    setosa   0.04076866  0.03340766   0.08908708   0.4008919
#>   4:    setosa   0.06794777  0.03340766   0.06681531   0.3674842
#>   5:    setosa   0.05435822  0.03340766   0.15590239   0.5345225
#>  ---                                                            
#> 146: virginica   0.57076130  0.73496842   0.53452248   0.3340766
#> 147: virginica   0.54358219  0.60133779   0.44543540   0.1670383
#> 148: virginica   0.57076130  0.63474545   0.48997894   0.3340766
#> 149: virginica   0.59794041  0.73496842   0.42316363   0.4677072
#> 150: virginica   0.55717174  0.56793014   0.35634832   0.3340766

pobk = po("splines", Boundary.knots = list(
  Petal.Length = c(0, 4), Petal.Width = c(4, 7), Sepal.Length = c(1, 5), Sepal.Width = c(3, 6))
)
pobk$train(list(task))[[1]]$data()
#>        Species Petal.Length Petal.Width Sepal.Length Sepal.Width
#>         <fctr>        <num>       <num>        <num>       <num>
#>   1:    setosa    0.2806243  -1.0155927    0.8218283  0.13363062
#>   2:    setosa    0.2806243  -1.0155927    0.7817391  0.00000000
#>   3:    setosa    0.2605797  -1.0155927    0.7416499  0.05345225
#>   4:    setosa    0.3006689  -1.0155927    0.7216054  0.02672612
#>   5:    setosa    0.2806243  -1.0155927    0.8017837  0.16035675
#>  ---                                                            
#> 146: virginica    1.0423188  -0.4543441    1.1425418  0.00000000
#> 147: virginica    1.0022297  -0.5612486    1.0623634 -0.13363062
#> 148: virginica    1.0423188  -0.5345225    1.1024526  0.00000000
#> 149: virginica    1.0824080  -0.4543441    1.0423188  0.10690450
#> 150: virginica    1.0222743  -0.5879747    0.9821851  0.00000000