PipeOpTargetTrafoSimple

Allows for target transformation operations of Task that have to be inverted later, where the transformation function is simply given by a function of the target. Typically this will be applied to a TaskRegr.

Format

R6Class object inheriting from PipeOpTargetTrafo/PipeOp

Construction

PipeOpTargetTrafoSimple$new(id = "targettrafosimple",
  param_vals = list())

id :: character(1)
Identifier of resulting object, default "targettrafosimple".

param_vals :: named list
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default list().

Input and Output Channels

Input and output channels are inherited from PipeOpTargetTrafo.

State

The $state is left empty (list()).

Parameters

The parameters are the parameters inherited from PipeOpTargetTrafo, as well as:

trafo :: function
Transformation function for the target. Should only be a function of the target, i.e., taking a single argument. Default is identity.

inverter :: function
Inversion of the transformation function for the target. Should only be a function of the target, i.e., taking a single argument. Default is identity.

new_target_name :: character(1)
Optionally give the transformed target a new name. By default the original name is used.

new_task_type :: character(1)
Optionally a new task type can be set. Legal types are listed in mlr_reflections$task_types$type.

Internals

Overloads PipeOpTargetTrafo's set_state(), train_target(), train_invert() and inverter(). Should be used in combination with PipeOpTargetInverter.

Methods

Only methods inherited from PipeOpTargetTrafo/PipeOp.

See also

Examples

library(mlr3)
task = tsk("boston_housing")
po = PipeOpTargetTrafoSimple$new("logtrafo", param_vals = list(
  trafo = function(x) log(x, base = 2),
  inverter = function(x) 2 ^ x)
)

po$train(list(task))

#> $fun
#> NULL
#> 
#> $output
#> <TaskRegr:boston_housing> (506 x 19)
#> * Target: medv
#> * Properties: -
#> * Features (18):
#>   - dbl (13): age, b, cmedv, crim, dis, indus, lat, lon, lstat, nox,
#>     ptratio, rm, zn
#>   - int (3): rad, tax, tract
#>   - fct (2): chas, town
#> 

po$predict(list(task))

#> $fun
#> function (inputs) 
#> {
#>     assert_list(inputs, len = 1L, types = "Prediction")
#>     self$inverter(inputs[[1L]], predict_phase_control)
#> }
#> <environment: 0x7fc5a2a99998>
#> 
#> $output
#> <TaskRegr:boston_housing> (506 x 19)
#> * Target: medv
#> * Properties: -
#> * Features (18):
#>   - dbl (13): age, b, cmedv, crim, dis, indus, lat, lon, lstat, nox,
#>     ptratio, rm, zn
#>   - int (3): rad, tax, tract
#>   - fct (2): chas, town
#> 

g = Graph$new()
g$add_pipeop(po)
g$add_pipeop(LearnerRegrRpart$new())
g$add_pipeop(PipeOpTargetInverter$new())
g$add_edge(src_id = "logtrafo", dst_id = "targetinverter",
  src_channel = 1, dst_channel = 1)
g$add_edge(src_id = "logtrafo", dst_id = "regr.rpart",
  src_channel = 2, dst_channel = 1)
g$add_edge(src_id = "regr.rpart", dst_id = "targetinverter",
  src_channel = 1, dst_channel = 2)

g$train(task)

#> $targetinverter.output
#> NULL
#> 

g$predict(task)

#> $targetinverter.output
#> <PredictionRegr> for 506 observations:
#>     row_id truth response
#>          1  24.0 24.35119
#>          2  21.6 20.48150
#>          3  34.7 32.83886
#> ---                      
#>        504  23.9 24.35119
#>        505  22.0 20.48150
#>        506  11.9 16.99570
#> 

#syntactic sugar using ppl():
tt = ppl("targettrafo", graph = PipeOpLearner$new(LearnerRegrRpart$new()))
tt$param_set$values$targettrafosimple.trafo = function(x) log(x, base = 2)
tt$param_set$values$targettrafosimple.inverter = function(x) 2 ^ x