Yeo-Johnson Transformation of Numeric Features

Conducts a Yeo-Johnson transformation on numeric features. It therefore estimates the optimal value of lambda for the transformation. See bestNormalize::yeojohnson() for details.

Format

R6Class object inheriting from PipeOpTaskPreproc/PipeOp.

Construction

PipeOpYeoJohnson$new(id = "yeojohnson", param_vals = list())

id :: character(1)
Identifier of resulting object, default "yeojohnson".

param_vals :: named list
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default list().

Input and Output Channels

Input and output channels are inherited from PipeOpTaskPreproc.

The output is the input Task with all affected numeric features replaced by their transformed versions.

State

The $state is a named list with the $state elements inherited from PipeOpTaskPreproc, as well as a list of class yeojohnson for each column, which is transformed.

Parameters

The parameters are the parameters inherited from PipeOpTaskPreproc, as well as:

eps :: numeric(1)
Tolerance parameter to identify the lambda parameter as zero. For details see yeojohnson().

standardize :: logical
Whether to center and scale the transformed values to attempt a standard normal distribution. For details see yeojohnson().

lower :: numeric(1)
Lower value for estimation of lambda parameter. For details see yeojohnson().

upper :: numeric(1)
Upper value for estimation of lambda parameter. For details see yeojohnson().

Internals

Uses the bestNormalize::yeojohnson function.

Methods

Only methods inherited from PipeOpTaskPreproc/PipeOp.

See also

Examples

 if (requireNamespace("bestNormalize")) { 
library("mlr3")

task = tsk("iris")
pop = po("yeojohnson")

task$data()
pop$train(list(task))[[1]]$data()

pop$state
 } 
#> $bc
#> $bc$Petal.Length
#> Standardized Yeo-Johnson Transformation with 150 nonmissing obs.:
#>  Estimated statistics:
#>  - lambda = 1.093219 
#>  - mean (before standardization) = 4.156174 
#>  - sd (before standardization) = 2.024973 
#> 
#> $bc$Petal.Width
#> Standardized Yeo-Johnson Transformation with 150 nonmissing obs.:
#>  Estimated statistics:
#>  - lambda = 0.8404896 
#>  - mean (before standardization) = 1.098253 
#>  - sd (before standardization) = 0.6787234 
#> 
#> $bc$Sepal.Length
#> Standardized Yeo-Johnson Transformation with 150 nonmissing obs.:
#>  Estimated statistics:
#>  - lambda = -0.3212232 
#>  - mean (before standardization) = 1.429598 
#>  - sd (before standardization) = 0.06498438 
#> 
#> $bc$Sepal.Width
#> Standardized Yeo-Johnson Transformation with 150 nonmissing obs.:
#>  Estimated statistics:
#>  - lambda = 0.03907448 
#>  - mean (before standardization) = 1.433769 
#>  - sd (before standardization) = 0.1131792 
#> 
#> 
#> $dt_columns
#> [1] "Petal.Length" "Petal.Width"  "Sepal.Length" "Sepal.Width" 
#> 
#> $affected_cols
#> [1] "Petal.Length" "Petal.Width"  "Sepal.Length" "Sepal.Width" 
#> 
#> $intasklayout
#> Key: <id>
#>              id    type
#>          <char>  <char>
#> 1: Petal.Length numeric
#> 2:  Petal.Width numeric
#> 3: Sepal.Length numeric
#> 4:  Sepal.Width numeric
#> 
#> $outtasklayout
#> Key: <id>
#>              id    type
#>          <char>  <char>
#> 1: Petal.Length numeric
#> 2:  Petal.Width numeric
#> 3: Sepal.Length numeric
#> 4:  Sepal.Width numeric
#> 
#> $outtaskshell
#> Empty data.table (0 rows and 5 cols): Species,Petal.Length,Petal.Width,Sepal.Length,Sepal.Width
#>