Create A Graph to Perform "One vs. Rest" classification.

Create a new Graph for a classification Task to perform "One vs. Rest" classification.

pipeline_ovr(graph)

Arguments

graph
Graph
Graph being wrapped between PipeOpOVRSplit and PipeOpOVRUnite. The Graph should return NULL during training and a classification Prediction during prediction.

Value

Graph

Examples

library("mlr3")

task = tsk("wine")

learner = lrn("classif.rpart")
learner$predict_type = "prob"

# Simple OVR
g1 = pipeline_ovr(learner)
g1$train(task)

#> $ovrunite.output
#> NULL
#> 

g1$predict(task)

#> $ovrunite.output
#> <PredictionClassif> for 178 observations:
#>     row_id truth response     prob.1     prob.2     prob.3
#>          1     1        1 0.93301153 0.04437722 0.02261125
#>          2     1        1 0.93301153 0.04437722 0.02261125
#>          3     1        1 0.93301153 0.04437722 0.02261125
#> ---                                                       
#>        176     3        3 0.08896784 0.04157376 0.86945841
#>        177     3        3 0.08896784 0.04157376 0.86945841
#>        178     3        3 0.01729144 0.04484461 0.93786395
#> 

# Bagged Learners
gr = po("replicate", reps = 3) %>>%
  po("subsample") %>>%
  learner %>>%
  po("classifavg", collect_multiplicity = TRUE)
g2 = pipeline_ovr(gr)
g2$train(task)

#> $ovrunite.output
#> NULL
#> 

g2$predict(task)

#> $ovrunite.output
#> <PredictionClassif> for 178 observations:
#>     row_id truth response     prob.1     prob.2     prob.3
#>          1     1        1 0.94252715 0.04134153 0.01613132
#>          2     1        1 0.94252715 0.04134153 0.01613132
#>          3     1        1 0.94252715 0.04134153 0.01613132
#> ---                                                       
#>        176     3        3 0.05608258 0.03979826 0.90411916
#>        177     3        3 0.27135681 0.03072168 0.69792150
#>        178     3        3 0.01269962 0.04162741 0.94567297
#> 

# Bagging outside OVR
g3 = po("replicate", reps = 3) %>>%
  pipeline_ovr(po("subsample") %>>% learner) %>>%
  po("classifavg", collect_multiplicity = TRUE)
g3$train(task)

#> $classifavg.output
#> NULL
#> 

g3$predict(task)

#> $classifavg.output
#> <PredictionClassif> for 178 observations:
#>     row_id truth response     prob.1     prob.2    prob.3
#>          1     1        1 0.95882795 0.02493035 0.0162417
#>          2     1        1 0.95882795 0.02493035 0.0162417
#>          3     1        1 0.95882795 0.02493035 0.0162417
#> ---                                                      
#>        176     3        3 0.13587563 0.02183435 0.8422900
#>        177     3        3 0.13587563 0.02183435 0.8422900
#>        178     3        3 0.03608536 0.02380071 0.9401139
#>