Shorthand PipeOp Constructor

Create

a PipeOp from mlr_pipeops from given ID

a PipeOpLearner from a Learner object

a PipeOpFilter from a Filter object

a PipeOpSelect from a Selector object

The object is initialized with given parameters and param_vals.

po(.obj, ...)

Arguments

.obj
[any]
The object from which to construct a PipeOp. If this is a character(1), it is looked up in the mlr_pipeops dictionary. Otherwise, it is converted to a PipeOp.

...
any
Additional parameters to give to constructed object. This may be an argument of the constructor of the PipeOp, in which case it is given to this constructor; or it may be a parameter value, in which case it is given to the param_vals argument of the constructor.

Examples

library("mlr3")

po("learner", lrn("classif.rpart"), cp = 0.3)

#> PipeOp: <classif.rpart> (not trained)
#> values: <xval=0, cp=0.3>
#> Input channels <name [train type, predict type]>:
#>   input [TaskClassif,TaskClassif]
#> Output channels <name [train type, predict type]>:
#>   output [NULL,PredictionClassif]

po(lrn("classif.rpart"), cp = 0.3)

#> PipeOp: <classif.rpart> (not trained)
#> values: <xval=0, cp=0.3>
#> Input channels <name [train type, predict type]>:
#>   input [TaskClassif,TaskClassif]
#> Output channels <name [train type, predict type]>:
#>   output [NULL,PredictionClassif]

# is equivalent with:
mlr_pipeops$get("learner", lrn("classif.rpart"),
  param_vals = list(cp = 0.3))

#> PipeOp: <classif.rpart> (not trained)
#> values: <xval=0, cp=0.3>
#> Input channels <name [train type, predict type]>:
#>   input [TaskClassif,TaskClassif]
#> Output channels <name [train type, predict type]>:
#>   output [NULL,PredictionClassif]