Shorthand Graph Constructor

Creates a Graph from mlr_graphs from given ID

ppl() taks a character(1) and returns a Graph. ppls() takes a character vector of any list and returns a list of possibly muliple Graphs.

Usage

ppl(.key, ...)

ppls(.keys, ...)

Arguments

.key

[character(1)]
The key of the Graph in mlr_graphs.

...

any
Additional parameters to give to constructed object. This may be an argument of the constructor of the underlying function.

.keys

[character]
The key of possibly multiple Graphs in mlr_graphs. If this is named, a named list is returned, but unlike pos() it will not set any $id slots.

Value

Graph (for ppl()) or list of Graphs (for ppls()).

Examples

 if (requireNamespace("rpart")) { 
library("mlr3")

gr = ppl("bagging", graph = po(lrn("regr.rpart")),
  averager = po("regravg", collect_multiplicity = TRUE))
 }