Robustify a learner

Creates a Graph that can be used to robustify any subsequent learner. Performs the following steps:

Drops empty factor levels using PipeOpFixFactors

Imputes numeric features using PipeOpImputeHist and PipeOpMissInd

Imputes factor features using PipeOpImputeNewlvl

Encodes factors using one-hot-encoding. Factors with a cardinality > max_cardinality are collapsed using [PipeOpCollapseFactors`].

If scaling, numeric features are scaled to mean 0 and standard deviation 1.

The graph is built conservatively, i.e. the function always tries to assure everything works. If a learner is provided, some steps can be left out, i.e. if the learner can deal with factor variables, no encoding is performed.

robustify_pipeline(task = NULL, learner = NULL,
  impute_missings = NULL, factors_to_numeric = NULL,
  max_cardinality = 1000)

Arguments

task
Task
A task to create a robustifying pipeline for. Optional, if omitted, the full pipeline is created.

learner
Learner
A learner to create a robustifying pipeline for. Optional, if omitted, a more conservative pipeline is built.

impute_missings
logical | NULL
Should missing values be imputed? Defaults to NULL, i.e imputes if the task has missing values and the learner can not handle them.

factors_to_numeric
logical | NULL
Should factors be encoded? Defaults to NULL, i.e encodes if the task has factors and the learner can not handle factors.

max_cardinality
integer
Maximum number of factor levels allowed. See above. Default: 1000.

Value

Graph

Examples

library(mlr3)
lrn = lrn("regr.rpart")
task = mlr_tasks$get("boston_housing")
gr = robustify_pipeline(task, lrn) %>>% po("learner", lrn)
resample(task, GraphLearner$new(gr), rsmp("holdout"))

#> <ResampleResult> of 1 iterations
#> * Task: boston_housing
#> * Learner: fixfactors.encode.removeconstants.regr.rpart
#> * Warnings: 0 in 0 iterations
#> * Errors: 0 in 0 iterations