Transform and Re-Transform the Target Variable

Wraps a Graph that transforms a target during training and inverts the transformation during prediction. This is done as follows:

Specify a transformation and inversion function using any subclass of PipeOpTargetTrafo, defaults to PipeOpTargetTrafoSimple, afterwards apply graph.

At the very end, during prediction the transformation is inverted using PipeOpTargetInverter.

To set a transformation and inversion function for PipeOpTargetTrafoSimple see the parameters trafo and inverter of the param_set of the resulting Graph.

pipeline_targettrafo(
  graph,
  trafo_pipeop = PipeOpTargetTrafoSimple$new(),
  id_prefix = ""
)

Arguments

graph
PipeOpLearner | Graph
A PipeOpLearner or Graph to create a robustifying pipeline for. If this is a Graph, the last PipeOp should be a PipeOpLearner and the first PipeOp should accept a single Task as input.

trafo_pipeop
PipeOp
A PipeOp that is a subclass of PipeOpTargetTrafo. Default is PipeOpTargetTrafoSimple.

id_prefix
character(1)
Optional id prefix to prepend to PipeOpTargetInverter ID. The resulting ID will be "[id_prefix]targetinverter". Default is "".

Value

Graph

Examples

library("mlr3")

tt = pipeline_targettrafo(PipeOpLearner$new(LearnerRegrRpart$new()))
tt$param_set$values$targettrafosimple.trafo = function(x) log(x, base = 2)
tt$param_set$values$targettrafosimple.inverter = function(x) 2 ^ x

# gives the same as
g = Graph$new()
g$add_pipeop(PipeOpTargetTrafoSimple$new(param_vals = list(
  trafo = function(x) log(x, base = 2),
  inverter = function(x) 2 ^ x)
  )
)
g$add_pipeop(LearnerRegrRpart$new())
g$add_pipeop(PipeOpTargetInverter$new())
g$add_edge(src_id = "targettrafosimple", dst_id = "targetinverter",
  src_channel = 1, dst_channel = 1)
g$add_edge(src_id = "targettrafosimple", dst_id = "regr.rpart",
  src_channel = 2, dst_channel = 1)
g$add_edge(src_id = "regr.rpart", dst_id = "targetinverter",
  src_channel = 1, dst_channel = 2)