Optimized Weighted Average of Features for Classification and Regression

Computes a weighted average of inputs. Used in the context of computing weighted averages of predictions.

Predictions are averaged using weights (in order of appearance in the data) which are optimized using nonlinear optimization from the package "nloptr" for a measure provided in measure (defaults to classif.acc for LearnerClassifAvg and regr.mse for LearnerRegrAvg). Learned weights can be obtained from $model. Using non-linear optimization is implemented in the SuperLearner R package. For a more detailed analysis the reader is referred to LeDell, 2015: Scalable Ensemble Learning and Computationally Efficient Variance Estimation.

mlr_learners_classif.avg

mlr_learners_regr.avg

Format

R6Class object inheriting from mlr3::LearnerClassif/mlr3::Learner.

Parameter Set

measure :: character(1) | Measure
Measure to optimized weights for. The Measure is either obtained from mlr_measures or directly supplied. Defaults to classif.acc for LearnerClassifAvg and regr.mse for LearnerRegrAvg

algorithm :: character(1)
Several nonlinear optimization methods from nloptr are available. See nloptr::nloptr.print.options() for a list of possible options. Note that we only allow for derivative free local or global algorithms, i.e. NLOPT_(G|L)N_.

Methods

LearnerClassifAvg$new(), id = "classif.avg")
(chr) -> self
Constructor.

LearnerRegrAvg$new(), id = "regr.avg")
(chr) -> self
Constructor.

See also