Dictionary of PipeOps

A simple Dictionary storing objects of class PipeOp. Each PipeOp has an associated help page, see mlr_pipeops_[id].

Format

R6Class object inheriting from mlr3misc::Dictionary.

Fields

Fields inherited from Dictionary, as well as:

metainf :: environment
Environment that stores the metainf argument of the $add() method. Only for internal use.

Methods

Methods inherited from Dictionary, as well as:

add(key, value, metainf = NULL)
(character(1), R6ClassGenerator, NULL | list)
Adds constructor value to the dictionary with key key, potentially overwriting a previously stored item. If metainf is not NULL (the default), it must be a list of arguments that will be given to the value constructor (i.e. value$new()) when it needs to be constructed for as.data.table PipeOp listing.

S3 methods

as.data.table(dict)
Dictionary -> data.table::data.table
Returns a data.table with the following columns:

key :: (character)
Key with which the PipeOp was registered to the Dictionary using the $add() method.

label :: (character)
Description of the PipeOp's functionality.

packages :: (character)
Set of all required packages for the PipeOp's train and predict methods.

tags :: (character)
A set of tags associated with the PipeOp describing its purpose.

feature_types :: (character)
Feature types the PipeOp operates on. Is NA for PipeOps that do not directly operate on a Task.

input.num, output.num :: (integer)
Number of the PipeOp's input and output channels. Is NA for PipeOps which accept a varying number of input and/or output channels depending a construction argument. See input and output fields of PipeOp.

input.type.train, input.type.predict, output.type.train, output.type.predict :: (character)
Types that are allowed as input to or returned as output of the PipeOp's $train() and $predict() methods.
A value of NULL means that a null object, e.g. no data, is taken as input or being returned as output. A value of "*" means that any type is possible.
If both input.type.train and output.type.train or both input.type.predict and output.type.predict contain values enclosed by square brackets ("[", "]"), then the respective input or channel is Multiplicity-aware. For more information, see Multiplicity.

See also

Examples

library("mlr3")

mlr_pipeops$get("learner", lrn("classif.rpart"))
#> PipeOp: <classif.rpart> (not trained)
#> values: <xval=0>
#> Input channels <name [train type, predict type]>:
#>   input [TaskClassif,TaskClassif]
#> Output channels <name [train type, predict type]>:
#>   output [NULL,PredictionClassif]

# equivalent:
po("learner", learner = lrn("classif.rpart"))
#> PipeOp: <classif.rpart> (not trained)
#> values: <xval=0>
#> Input channels <name [train type, predict type]>:
#>   input [TaskClassif,TaskClassif]
#> Output channels <name [train type, predict type]>:
#>   output [NULL,PredictionClassif]

# all PipeOps currently in the dictionary:
as.data.table(mlr_pipeops)[, c("key", "input.num", "output.num", "packages")]
#> Key: <key>
#>                       key input.num output.num                         packages
#>                    <char>     <int>      <int>                           <list>
#>  1:                  adas         1          1        mlr3pipelines,smotefamily
#>  2:               blsmote         1          1        mlr3pipelines,smotefamily
#>  3:                boxcox         1          1      mlr3pipelines,bestNormalize
#>  4:                branch         1         NA                    mlr3pipelines
#>  5:                 chunk         1         NA                    mlr3pipelines
#>  6:        classbalancing         1          1                    mlr3pipelines
#>  7:            classifavg        NA          1              mlr3pipelines,stats
#>  8:          classweights         1          1                    mlr3pipelines
#>  9:              colapply         1          1                    mlr3pipelines
#> 10:       collapsefactors         1          1                    mlr3pipelines
#> 11:              colroles         1          1                    mlr3pipelines
#> 12:                  copy         1         NA                    mlr3pipelines
#> 13:          datefeatures         1          1                    mlr3pipelines
#> 14:                decode         1          1                    mlr3pipelines
#> 15:                encode         1          1              mlr3pipelines,stats
#> 16:          encodeimpact         1          1                    mlr3pipelines
#> 17:            encodelmer         1          1        mlr3pipelines,lme4,nloptr
#> 18:     encodeplquantiles         1          1              mlr3pipelines,stats
#> 19:          encodepltree         1          1         mlr3pipelines,mlr3,rpart
#> 20:          featureunion        NA          1                    mlr3pipelines
#> 21:                filter         1          1                    mlr3pipelines
#> 22:            fixfactors         1          1                    mlr3pipelines
#> 23:               histbin         1          1           mlr3pipelines,graphics
#> 24:                   ica         1          1            mlr3pipelines,fastICA
#> 25:        imputeconstant         1          1                    mlr3pipelines
#> 26:            imputehist         1          1           mlr3pipelines,graphics
#> 27:         imputelearner         1          1                    mlr3pipelines
#> 28:            imputemean         1          1                    mlr3pipelines
#> 29:          imputemedian         1          1              mlr3pipelines,stats
#> 30:            imputemode         1          1                    mlr3pipelines
#> 31:             imputeoor         1          1                    mlr3pipelines
#> 32:          imputesample         1          1                    mlr3pipelines
#> 33:             kernelpca         1          1            mlr3pipelines,kernlab
#> 34:               learner         1          1                    mlr3pipelines
#> 35:            learner_cv         1          1                    mlr3pipelines
#> 36:     learner_pi_cvplus         1          1                    mlr3pipelines
#> 37:     learner_quantiles         1          1                    mlr3pipelines
#> 38:               missind         1          1                    mlr3pipelines
#> 39:           modelmatrix         1          1              mlr3pipelines,stats
#> 40:     multiplicityexply         1         NA                    mlr3pipelines
#> 41:     multiplicityimply        NA          1                    mlr3pipelines
#> 42:                mutate         1          1                    mlr3pipelines
#> 43:              nearmiss         1          1             mlr3pipelines,themis
#> 44:                   nmf         1          1           mlr3pipelines,MASS,NMF
#> 45:                   nop         1          1                    mlr3pipelines
#> 46:              ovrsplit         1          1                    mlr3pipelines
#> 47:              ovrunite         1          1                    mlr3pipelines
#> 48:                   pca         1          1                    mlr3pipelines
#> 49:                 proxy        NA          1                    mlr3pipelines
#> 50:           quantilebin         1          1              mlr3pipelines,stats
#> 51:      randomprojection         1          1                    mlr3pipelines
#> 52:        randomresponse         1          1                    mlr3pipelines
#> 53:               regravg        NA          1                    mlr3pipelines
#> 54:       removeconstants         1          1                    mlr3pipelines
#> 55:         renamecolumns         1          1                    mlr3pipelines
#> 56:             replicate         1          1                    mlr3pipelines
#> 57:              rowapply         1          1                    mlr3pipelines
#> 58:                 scale         1          1                    mlr3pipelines
#> 59:           scalemaxabs         1          1                    mlr3pipelines
#> 60:            scalerange         1          1                    mlr3pipelines
#> 61:                select         1          1                    mlr3pipelines
#> 62:                 smote         1          1        mlr3pipelines,smotefamily
#> 63:               smotenc         1          1             mlr3pipelines,themis
#> 64:           spatialsign         1          1                    mlr3pipelines
#> 65:             subsample         1          1                    mlr3pipelines
#> 66:          targetinvert         2          1                    mlr3pipelines
#> 67:          targetmutate         1          2                    mlr3pipelines
#> 68: targettrafoscalerange         1          2                    mlr3pipelines
#> 69:        textvectorizer         1          1 mlr3pipelines,quanteda,stopwords
#> 70:             threshold         1          1                    mlr3pipelines
#> 71:                 tomek         1          1             mlr3pipelines,themis
#> 72:         tunethreshold         1          1              mlr3pipelines,bbotk
#> 73:              unbranch        NA          1                    mlr3pipelines
#> 74:                vtreat         1          1             mlr3pipelines,vtreat
#> 75:            yeojohnson         1          1      mlr3pipelines,bestNormalize
#>                       key input.num output.num                         packages