Dictionary of PipeOps

A simple Dictionary storing objects of class PipeOp. Each PipeOp has an associated help page, see mlr_pipeops_[id].

Format

R6Class object inheriting from mlr3misc::Dictionary.

Fields

Fields inherited from Dictionary, as well as:

metainf :: environment
Environment that stores the metainf argument of the $add() method. Only for internal use.

Methods

Methods inherited from Dictionary, as well as:

add(key, value, metainf = NULL)
(character(1), R6ClassGenerator, NULL | list)
Adds constructor value to the dictionary with key key, potentially overwriting a previously stored item. If metainf is not NULL (the default), it must be a list of arguments that will be given to the value constructor (i.e. value$new()) when it needs to be constructed for as.data.table PipeOp listing.

S3 methods

as.data.table(dict)
Dictionary -> data.table::data.table
Returns a data.table with columns key (character), packages (character), input.num (integer), output.num (integer), input.type.train (character), input.type.predict (character), output.type.train (character), output.type.predict (character).

See also

Examples

library("mlr3")

mlr_pipeops$get("learner", lrn("classif.rpart"))

#> PipeOp: <classif.rpart> (not trained)
#> values: <xval=0>
#> Input channels <name [train type, predict type]>:
#>   input [TaskClassif,TaskClassif]
#> Output channels <name [train type, predict type]>:
#>   output [NULL,PredictionClassif]

# equivalent:
po("learner", learner = lrn("classif.rpart"))

#> PipeOp: <classif.rpart> (not trained)
#> values: <xval=0>
#> Input channels <name [train type, predict type]>:
#>   input [TaskClassif,TaskClassif]
#> Output channels <name [train type, predict type]>:
#>   output [NULL,PredictionClassif]

# all PipeOps currently in the dictionary:
as.data.table(mlr_pipeops)[, c("key", "input.num", "output.num", "packages")]

#>                 key input.num output.num      packages
#>  1:          boxcox         1          1 bestNormalize
#>  2:          branch         1         NA              
#>  3:           chunk         1         NA              
#>  4:  classbalancing         1          1              
#>  5:      classifavg        NA          1         stats
#>  6:    classweights         1          1              
#>  7:        colapply         1          1              
#>  8: collapsefactors         1          1              
#>  9:            copy         1         NA              
#> 10:          encode         1          1         stats
#> 11:    encodeimpact         1          1              
#> 12:      encodelmer         1          1   lme4,nloptr
#> 13:    featureunion        NA          1              
#> 14:          filter         1          1              
#> 15:      fixfactors         1          1              
#> 16:         histbin         1          1      graphics
#> 17:             ica         1          1       fastICA
#> 18:      imputehist         1          1      graphics
#> 19:      imputemean         1          1              
#> 20:    imputemedian         1          1         stats
#> 21:    imputenewlvl         1          1              
#> 22:    imputesample         1          1              
#> 23:       kernelpca         1          1       kernlab
#> 24:         learner         1          1              
#> 25:      learner_cv         1          1              
#> 26:         missind         1          1              
#> 27:     modelmatrix         1          1         stats
#> 28:          mutate         1          1              
#> 29:             nop         1          1              
#> 30:             pca         1          1              
#> 31:     quantilebin         1          1         stats
#> 32:         regravg        NA          1              
#> 33: removeconstants         1          1              
#> 34:           scale         1          1              
#> 35:     scalemaxabs         1          1              
#> 36:      scalerange         1          1              
#> 37:          select         1          1              
#> 38:           smote         1          1   smotefamily
#> 39:     spatialsign         1          1              
#> 40:       subsample         1          1              
#> 41:        unbranch        NA          1              
#> 42:      yeojohnson         1          1 bestNormalize
#>                 key input.num output.num      packages