Dictionary of PipeOps

A simple Dictionary storing objects of class PipeOp. Each PipeOp has an associated help page, see mlr_pipeops_[id].

Format

R6Class object inheriting from mlr3misc::Dictionary.

Fields

Fields inherited from Dictionary, as well as:

metainf :: environment
Environment that stores the metainf argument of the $add() method. Only for internal use.

Methods

Methods inherited from Dictionary, as well as:

add(key, value, metainf = NULL)
(character(1), R6ClassGenerator, NULL | list)
Adds constructor value to the dictionary with key key, potentially overwriting a previously stored item. If metainf is not NULL (the default), it must be a list of arguments that will be given to the value constructor (i.e. value$new()) when it needs to be constructed for as.data.table PipeOp listing.

S3 methods

as.data.table(dict)
Dictionary -> data.table::data.table
Returns a data.table with columns key (character), packages (character), input.num (integer), output.num (integer), input.type.train (character), input.type.predict (character), output.type.train (character), output.type.predict (character).

See also

Examples

library("mlr3")

mlr_pipeops$get("learner", lrn("classif.rpart"))
#> PipeOp: <classif.rpart> (not trained)
#> values: <xval=0>
#> Input channels <name [train type, predict type]>:
#>   input [TaskClassif,TaskClassif]
#> Output channels <name [train type, predict type]>:
#>   output [NULL,PredictionClassif]

# equivalent:
po("learner", learner = lrn("classif.rpart"))
#> PipeOp: <classif.rpart> (not trained)
#> values: <xval=0>
#> Input channels <name [train type, predict type]>:
#>   input [TaskClassif,TaskClassif]
#> Output channels <name [train type, predict type]>:
#>   output [NULL,PredictionClassif]

# all PipeOps currently in the dictionary:
as.data.table(mlr_pipeops)[, c("key", "input.num", "output.num", "packages")]
#>                       key input.num output.num                         packages
#>  1:                boxcox         1          1      mlr3pipelines,bestNormalize
#>  2:                branch         1         NA                    mlr3pipelines
#>  3:                 chunk         1         NA                    mlr3pipelines
#>  4:        classbalancing         1          1                    mlr3pipelines
#>  5:            classifavg        NA          1              mlr3pipelines,stats
#>  6:          classweights         1          1                    mlr3pipelines
#>  7:              colapply         1          1                    mlr3pipelines
#>  8:       collapsefactors         1          1                    mlr3pipelines
#>  9:              colroles         1          1                    mlr3pipelines
#> 10:                  copy         1         NA                    mlr3pipelines
#> 11:          datefeatures         1          1                    mlr3pipelines
#> 12:                encode         1          1              mlr3pipelines,stats
#> 13:          encodeimpact         1          1                    mlr3pipelines
#> 14:            encodelmer         1          1        mlr3pipelines,lme4,nloptr
#> 15:          featureunion        NA          1                    mlr3pipelines
#> 16:                filter         1          1                    mlr3pipelines
#> 17:            fixfactors         1          1                    mlr3pipelines
#> 18:               histbin         1          1           mlr3pipelines,graphics
#> 19:                   ica         1          1            mlr3pipelines,fastICA
#> 20:        imputeconstant         1          1                    mlr3pipelines
#> 21:            imputehist         1          1           mlr3pipelines,graphics
#> 22:         imputelearner         1          1                    mlr3pipelines
#> 23:            imputemean         1          1                    mlr3pipelines
#> 24:          imputemedian         1          1              mlr3pipelines,stats
#> 25:            imputemode         1          1                    mlr3pipelines
#> 26:             imputeoor         1          1                    mlr3pipelines
#> 27:          imputesample         1          1                    mlr3pipelines
#> 28:             kernelpca         1          1            mlr3pipelines,kernlab
#> 29:               learner         1          1                    mlr3pipelines
#> 30:            learner_cv         1          1                    mlr3pipelines
#> 31:               missind         1          1                    mlr3pipelines
#> 32:           modelmatrix         1          1              mlr3pipelines,stats
#> 33:     multiplicityexply         1         NA                    mlr3pipelines
#> 34:     multiplicityimply        NA          1                    mlr3pipelines
#> 35:                mutate         1          1                    mlr3pipelines
#> 36:                   nmf         1          1           mlr3pipelines,MASS,NMF
#> 37:                   nop         1          1                    mlr3pipelines
#> 38:              ovrsplit         1          1                    mlr3pipelines
#> 39:              ovrunite         1          1                    mlr3pipelines
#> 40:                   pca         1          1                    mlr3pipelines
#> 41:                 proxy        NA          1                    mlr3pipelines
#> 42:           quantilebin         1          1              mlr3pipelines,stats
#> 43:      randomprojection         1          1                    mlr3pipelines
#> 44:        randomresponse         1          1                    mlr3pipelines
#> 45:               regravg        NA          1                    mlr3pipelines
#> 46:       removeconstants         1          1                    mlr3pipelines
#> 47:         renamecolumns         1          1                    mlr3pipelines
#> 48:             replicate         1          1                    mlr3pipelines
#> 49:                 scale         1          1                    mlr3pipelines
#> 50:           scalemaxabs         1          1                    mlr3pipelines
#> 51:            scalerange         1          1                    mlr3pipelines
#> 52:                select         1          1                    mlr3pipelines
#> 53:                 smote         1          1        mlr3pipelines,smotefamily
#> 54:           spatialsign         1          1                    mlr3pipelines
#> 55:             subsample         1          1                    mlr3pipelines
#> 56:          targetinvert         2          1                    mlr3pipelines
#> 57:          targetmutate         1          2                    mlr3pipelines
#> 58: targettrafoscalerange         1          2                    mlr3pipelines
#> 59:        textvectorizer         1          1 mlr3pipelines,quanteda,stopwords
#> 60:             threshold         1          1                    mlr3pipelines
#> 61:         tunethreshold         1          1              mlr3pipelines,bbotk
#> 62:              unbranch        NA          1                    mlr3pipelines
#> 63:                vtreat         1          1             mlr3pipelines,vtreat
#> 64:            yeojohnson         1          1      mlr3pipelines,bestNormalize
#>                       key input.num output.num                         packages