PipeOpICA

Extracts statistically independent components from data. Only affects numerical features. See fastICA::fastICA for details.

Format

R6Class object inheriting from PipeOpTaskPreproc/PipeOp.

Construction

PipeOpICA$new(id = "ica", param_vals = list())

id :: character(1)
Identifier of resulting object, default "ica".

param_vals :: named list
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default list().

Input and Output Channels

Input and output channels are inherited from PipeOpTaskPreproc.

The output is the input Task with all affected numeric parameters replaced by independent components.

State

The $state is a named list with the $state elements inherited from PipeOpTaskPreproc, as well as the elements of the function fastICA::fastICA(), with the exception of the $X and $S slots. These are in particular:

K :: matrix
Matrix that projects data onto the first n.comp principal components. See fastICA().

W :: matrix
Estimated un-mixing matrix. See fastICA().

A :: matrix
Estimated mixing matrix. See fastICA().

center :: numeric
The mean of each numeric feature during training.

Parameters

The parameters are the parameters inherited from PipeOpTaskPreproc, as well as the following parameters based on fastICA():

n.comp :: numeric(1)
Number of components to extract. Default is NULL, which sets it to the number of available numeric columns.

alg.typ:: character(1)
Algorithm type. One of “parallel” (default) or “deflation”.

fun :: character(1)
One of “logcosh” (default) or “exp”.

alpha :: numeric(1)
In range [1, 2], Used for negentropy calculation when fun is “logcosh”. Default is 1.0.

method :: character(1)
Internal calculation method. “C” (default) or “R”. See fastICA().

row.norm :: logical(1)
Logical value indicating whether rows should be standardized beforehand. Default is FALSE.

maxit :: numeric(1)
Maximum number of iterations. Default is 200.

tol :: numeric(1)
Tolerance for convergence, default is 1e-4.

verbose logical(1)
Logical value indicating the level of output during the run of the algorithm. Default is FALSE.

w.init:: matrix
Initial un-mixing matrix. See fastICA(). Default is NULL.

Internals

Uses the fastICA() function.

Methods

Only methods inherited from PipeOpTaskPreproc/PipeOp.

See also

Examples

library("mlr3")

task = tsk("iris")
pop = po("ica")

task$data()

#>        Species Petal.Length Petal.Width Sepal.Length Sepal.Width
#>   1:    setosa          1.4         0.2          5.1         3.5
#>   2:    setosa          1.4         0.2          4.9         3.0
#>   3:    setosa          1.3         0.2          4.7         3.2
#>   4:    setosa          1.5         0.2          4.6         3.1
#>   5:    setosa          1.4         0.2          5.0         3.6
#>  ---                                                            
#> 146: virginica          5.2         2.3          6.7         3.0
#> 147: virginica          5.0         1.9          6.3         2.5
#> 148: virginica          5.2         2.0          6.5         3.0
#> 149: virginica          5.4         2.3          6.2         3.4
#> 150: virginica          5.1         1.8          5.9         3.0

pop$train(list(task))[[1]]$data()

#>        Species          V1         V2          V3         V4
#>   1:    setosa  0.02964015  1.3923648  0.37563489  0.2610234
#>   2:    setosa -0.07974739  1.3297421 -0.97165002  0.3867335
#>   3:    setosa -0.15804277  1.3493647 -0.34942811 -0.3569939
#>   4:    setosa  0.31308719  1.2047706 -0.38031792 -0.8801721
#>   5:    setosa  0.10772239  1.3726166  0.73658482 -0.2011008
#>  ---                                                        
#> 146: virginica -2.59293017 -0.8277007 -0.31819342  1.2708891
#> 147: virginica -1.05284491 -0.7561224 -1.36118379  0.7285023
#> 148: virginica -0.98406266 -0.8272547  0.09222046  0.3429698
#> 149: virginica -1.77153523 -1.0479842  1.39031431 -1.2507800
#> 150: virginica  0.04140495 -0.8754143  0.57112826 -1.5517294

pop$state

#> $K
#>            [,1]       [,2]       [,3]      [,4]
#> [1,] -0.4180098  0.3531217 -0.2735163 -3.118456
#> [2,] -0.1748261  0.1537381 -1.9583093  4.897992
#> [3,] -0.1763375 -1.3373258  2.0881803  2.050340
#> [4,]  0.0412425 -1.4871770 -2.1451574 -2.077869
#> 
#> $W
#>              [,1]        [,2]        [,3]       [,4]
#> [1,] -0.011944047  0.98686291 -0.06934104 -0.1454330
#> [2,]  0.004171424 -0.14222864 -0.79871639 -0.5846415
#> [3,]  0.597007954  0.06805965 -0.47797479  0.6406945
#> [4,] -0.802135550  0.03522063 -0.35886494  0.4759769
#> 
#> $A
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.06786320 -0.17492263  0.06541297 -0.06359578
#> [2,] -1.74875246 -0.73617361 -0.67229657  0.20885734
#> [3,]  0.09040458  0.05241777  0.21387398  0.37163002
#> [4,]  0.15680913  0.04284564  0.42322406  0.05420524
#> 
#> $center
#> Petal.Length  Petal.Width Sepal.Length  Sepal.Width 
#>     3.758000     1.199333     5.843333     3.057333 
#> 
#> $dt_columns
#> [1] "Petal.Length" "Petal.Width"  "Sepal.Length" "Sepal.Width" 
#> 
#> $affected_cols
#> [1] "Petal.Length" "Petal.Width"  "Sepal.Length" "Sepal.Width" 
#> 
#> $intasklayout
#>              id    type
#> 1: Petal.Length numeric
#> 2:  Petal.Width numeric
#> 3: Sepal.Length numeric
#> 4:  Sepal.Width numeric
#> 
#> $outtasklayout
#>    id    type
#> 1: V1 numeric
#> 2: V2 numeric
#> 3: V3 numeric
#> 4: V4 numeric
#>