Wrap a Learner into a PipeOp

Wraps an mlr3::Learner into a PipeOp.

Inherits the $param_set (and therefore $param_set$values) from the Learner it is constructed from.

Using PipeOpLearner, it is possible to embed mlr3::Learners into Graphs, which themselves can be turned into Learners using GraphLearner. This way, preprocessing and ensemble methods can be included into a machine learning pipeline which then can be handled as singular object for resampling, benchmarking and tuning.

Format

R6Class object inheriting from PipeOp.

Construction

PipeOpLearner$new(learner, id = NULL, param_vals = list())

learner :: Learner | character(1) Learner to wrap, or a string identifying a Learner in the mlr3::mlr_learners Dictionary. This argument is always cloned; to access the Learner inside PipeOpLearner by-reference, use $learner.

id :: character(1) Identifier of the resulting object, internally defaulting to the id of the Learner being wrapped.

param_vals :: named list
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default list().

Input and Output Channels

PipeOpLearner has one input channel named "input", taking a Task specific to the Learner type given to learner during construction; both during training and prediction.

PipeOpLearner has one output channel named "output", producing NULL during training and a Prediction subclass during prediction; this subclass is specific to the Learner type given to learner during construction.

The output during prediction is the Prediction on the prediction input data, produced by the Learner trained on the training input data.

State

The $state is set to the $state slot of the Learner object. It is a named list with members:

model :: any
Model created by the Learner's $.train() function.

train_log :: data.table with columns class (character), msg (character)
Errors logged during training.

train_time :: numeric(1)
Training time, in seconds.

predict_log :: NULL | data.table with columns class (character), msg (character)
Errors logged during prediction.

predict_time :: NULL | numeric(1) Prediction time, in seconds.

Parameters

The parameters are exactly the parameters of the Learner wrapped by this object.

Internals

The $state is currently not updated by prediction, so the $state$predict_log and $state$predict_time will always be NULL.

Fields

Fields inherited from PipeOp, as well as:

learner :: Learner
Learner that is being wrapped. Read-only.

learner_model :: Learner
Learner that is being wrapped. This learner contains the model if the PipeOp is trained. Read-only.

validate :: "predefined" or NULL
This field can only be set for Learners that have the "validation" property. Setting the field to "predefined" means that the wrapped Learner will use the internal validation task, otherwise it will be ignored. Note that specifying how the validation data is created is possible via the $validate field of the GraphLearner. For each PipeOp it is then only possible to either use it ("predefined") or not use it (NULL). Also see set_validate.GraphLearner for more information.

internal_tuned_values :: named list() or NULL
The internally tuned values if the wrapped Learner supports internal tuning, NULL otherwise.

internal_valid_scores :: named list() or NULL
The internal validation scores if the wrapped Learner supports internal validation, NULL otherwise.

Methods

Methods inherited from PipeOp.

See also

Examples