PipeOpNOP

Simply pushes the input forward. Can be useful during Graph construction using the %>>%-operator to specify which PipeOp gets connected to which.

Format

R6Class object inheriting from PipeOp.

Construction

PipeOpNOP$new(id = "nop", param_vals = list())

id :: character(1)
Identifier of resulting object, default "nop".

param_vals :: named list
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default list().

Input and Output Channels

PipeOpNOP has one input channel named "input", taking any input ("*") both during training and prediction.

PipeOpNOP has one output channel named "output", producing the object given as input ("*") without changes.

State

The $state is left empty (list()).

Parameters

PipeOpNOP has no parameters.

Internals

PipeOpNOP is a useful "default" stand-in for a PipeOp/Graph that does nothing.

Fields

Only fields inherited from PipeOp.

Methods

Only methods inherited from PipeOp.

See also

Examples

library("mlr3")

nop = po("nop")

nop$train(list(1))

#> $output
#> [1] 1
#> 

# use `gunion` and `%>>%` to create a "bypass"
# next to "pca"
gr = gunion(list(
  po("pca"),
  nop
)) %>>% po("featureunion")

gr$train(tsk("iris"))[[1]]$data()

#>        Species       PC1         PC2         PC3          PC4 Petal.Length
#>   1:    setosa -2.684126  0.31939725 -0.02791483 -0.002262437          1.4
#>   2:    setosa -2.714142 -0.17700123 -0.21046427 -0.099026550          1.4
#>   3:    setosa -2.888991 -0.14494943  0.01790026 -0.019968390          1.3
#>   4:    setosa -2.745343 -0.31829898  0.03155937  0.075575817          1.5
#>   5:    setosa -2.728717  0.32675451  0.09007924  0.061258593          1.4
#>  ---                                                                      
#> 146: virginica  1.944110  0.18753230  0.17782509 -0.426195940          5.2
#> 147: virginica  1.527167 -0.37531698 -0.12189817 -0.254367442          5.0
#> 148: virginica  1.764346  0.07885885  0.13048163 -0.137001274          5.2
#> 149: virginica  1.900942  0.11662796  0.72325156 -0.044595305          5.4
#> 150: virginica  1.390189 -0.28266094  0.36290965  0.155038628          5.1
#>      Petal.Width Sepal.Length Sepal.Width
#>   1:         0.2          5.1         3.5
#>   2:         0.2          4.9         3.0
#>   3:         0.2          4.7         3.2
#>   4:         0.2          4.6         3.1
#>   5:         0.2          5.0         3.6
#>  ---                                     
#> 146:         2.3          6.7         3.0
#> 147:         1.9          6.3         2.5
#> 148:         2.0          6.5         3.0
#> 149:         2.3          6.2         3.4
#> 150:         1.8          5.9         3.0