Bag-of-word Representation of Character Features

Computes a bag-of-word representation from a (set of) columns. Columns of type character are split up into words. Uses the quanteda::dfm(), quanteda::dfm_trim() from the 'quanteda' package. TF-IDF computation works similarly to quanteda::dfm_tfidf() but has been adjusted for train/test data split using quanteda::docfreq() and quanteda::dfm_weight()

In short:

Per default, produces a bag-of-words representation

If n is set to values > 1, ngrams are computed

If df_trim parameters are set, the bag-of-words is trimmed.

The scheme_tf parameter controls term-frequency (per-document, i.e. per-row) weighting

The scheme_df parameter controls the document-frequency (per token, i.e. per-column) weighting.

Parameters specify arguments to quanteda's dfm, dfm_trim, docfreq and dfm_weight. What belongs to what can be obtained from each params tags where tokenizer are arguments passed on to quanteda::dfm(). Defaults to a bag-of-words representation with token counts as matrix entries.

In order to perform the default dfm_tfidf weighting, set the scheme_df parameter to "inverse". The scheme_df parameter is initialized to "unary", which disables document frequency weighting.

The pipeop works as follows:

Words are tokenized using quanteda::tokens.

Ngrams are computed using quanteda::tokens_ngrams

A document-frequency matrix is computed using quanteda::dfm

The document-frequency matrix is trimmed using quanteda::dfm_trim during train-time.

The document-frequency matrix is re-weighted (similar to quanteda::dfm_tfidf) if scheme_df is not set to "unary".

Format

R6Class object inheriting from PipeOpTaskPreproc/PipeOp.

Construction

PipeOpTextVectorizer$new(id = "textvectorizer", param_vals = list())

id :: character(1)
Identifier of resulting object, default "textvectorizer".

param_vals :: named list
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default list().

Input and Output Channels

Input and output channels are inherited from PipeOpTaskPreproc.

The output is the input Task with all affected features converted to a bag-of-words representation.

State

The $state is a list with element 'cols': A vector of extracted columns.

Parameters

The parameters are the parameters inherited from PipeOpTaskPreproc, as well as:

return_type :: character(1)
Whether to return an integer representation ("integer-sequence") or a Bag-of-words ("bow"). If set to "integer_sequence", tokens are replaced by an integer and padded/truncated to sequence_length. If set to "factor_sequence", tokens are replaced by a factor and padded/truncated to sequence_length. If set to 'bow', a possibly weighted bag-of-words matrix is returned. Defaults to bow.

stopwords_language :: character(1)
Language to use for stopword filtering. Needs to be either "none", a language identifier listed in stopwords::stopwords_getlanguages("snowball") ("de", "en", ...) or "smart". "none" disables language-specific stopwords. "smart" coresponds to stopwords::stopwords(source = "smart"), which contains English stopwords and also removes one-character strings. Initialized to "smart".

extra_stopwords :: character
Extra stopwords to remove. Must be a character vector containing individual tokens to remove. Initialized to character(0). When n is set to values greater than 1, this can also contain stop-ngrams.

tolower :: logical(1)
Convert to lower case? See quanteda::dfm. Default: TRUE.

stem :: logical(1)
Perform stemming? See quanteda::dfm. Default: FALSE.

what :: character(1)
Tokenization splitter. See quanteda::tokens. Default: word.

remove_punct :: logical(1)
See quanteda::tokens. Default: FALSE.

remove_url :: logical(1)
See quanteda::tokens. Default: FALSE.

remove_symbols :: logical(1)
See quanteda::tokens. Default: FALSE.

remove_numbers :: logical(1)
See quanteda::tokens. Default: FALSE.

remove_separators :: logical(1)
See quanteda::tokens. Default: TRUE.

split_hypens :: logical(1)
See quanteda::tokens. Default: FALSE.

n :: integer
Vector of ngram lengths. See quanteda::tokens_ngrams. Initialized to 1, deviating from the base function's default. Note that this can be a vector of multiple values, to construct ngrams of multiple orders.

skip :: integer
Vector of skips. See quanteda::tokens_ngrams. Default: 0. Note that this can be a vector of multiple values.

sparsity :: numeric(1)
Desired sparsity of the 'tfm' matrix. See quanteda::dfm_trim. Default: NULL.

max_termfreq :: numeric(1)
Maximum term frequency in the 'tfm' matrix. See quanteda::dfm_trim. Default: NULL.

min_termfreq :: numeric(1)
Minimum term frequency in the 'tfm' matrix. See quanteda::dfm_trim. Default: NULL.

termfreq_type :: character(1)
How to asess term frequency. See quanteda::dfm_trim. Default: "count".

scheme_df :: character(1)
Weighting scheme for document frequency: See quanteda::docfreq. Initialized to "unary" (1 for each document, deviating from base function default).

smoothing_df :: numeric(1)
See quanteda::docfreq. Default: 0.

k_df :: numeric(1)
k parameter given to quanteda::docfreq (see there). Default is 0.

threshold_df :: numeric(1)
See quanteda::docfreq. Default: 0. Only considered for scheme_df = "count".

base_df :: numeric(1)
The base for logarithms in quanteda::docfreq (see there). Default: 10.

scheme_tf :: character(1)
Weighting scheme for term frequency: See quanteda::dfm_weight. Default: "count".

k_tf :: numeric(1)
k parameter given to quanteda::dfm_weight (see there). Default behaviour is 0.5.

base_df :: numeric(1)
The base for logarithms in quanteda::dfm_weight (see there). Default: 10.

#' * sequence_length :: integer(1)
The length of the integer sequence. Defaults to Inf, i.e. all texts are padded to the length of the longest text. Only relevant for "return_type" : "integer_sequence"

Internals

See Description. Internally uses the quanteda package. Calls quanteda::tokens, quanteda::tokens_ngrams and quanteda::dfm. During training, quanteda::dfm_trim is also called. Tokens not seen during training are dropped during prediction.

Methods

Only methods inherited from PipeOpTaskPreproc/PipeOp.

See also

Examples