PipeOpThreshold

Change the threshold of a Prediction during the predict step. The incoming Learner's $predict_type needs to be "prob". Internally calls PredictionClassif$set_threshold.

Format

R6Class inheriting from PipeOp.

Construction

PipeOpThreshold$new(id = "threshold", param_vals = list())

id :: character(1) Identifier of the resulting object, default "threshold".

param_vals :: named list
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Defaults to numeric(0).

Input and Output Channels

During training, the input and output are NULL. A PredictionClassif is required as input and returned as output during prediction.

State

The $state is left empty (list()).

Parameters

thresholds :: numeric
A numeric vector of thresholds for the different class levels. May have length 1 for binary classification predictions, must otherwise have length of the number of target classes; see PredictionClassif's $set_threshold() method. Initialized to 0.5, i.e. thresholding for binary classification at level 0.5.

Fields

Only fields inherited from PipeOp.

Methods

Only methods inherited from PipeOp.

See also

Examples

library("mlr3")
t = tsk("german_credit")
gr = po(lrn("classif.rpart", predict_type = "prob")) %>>%
  po("threshold", param_vals = list(thresholds = 0.9))
gr$train(t)

#> $threshold.output
#> NULL
#> 

gr$predict(t)

#> $threshold.output
#> <PredictionClassif> for 1000 observations:
#>     row_id truth response prob.good  prob.bad
#>          1  good      bad 0.8767123 0.1232877
#>          2   bad      bad 0.1388889 0.8611111
#>          3  good      bad 0.8687090 0.1312910
#> ---                                          
#>        998  good      bad 0.8687090 0.1312910
#>        999   bad      bad 0.3795620 0.6204380
#>       1000  good      bad 0.7391304 0.2608696
#>